refactor(timeline): split bot domains out of room_screen mod.rs
Stage 1 of the room_screen split, as verbatim code motion: the bot/agent
message rendering layer (bot_message.rs), the octos/agent-chat action
button + approval protocol (octos_actions.rs), and bot identity
detection / app-service admin helpers (bot_admin.rs) become submodules.

Moved items gain pub(super); the externally used is_known_or_likely_bot
is re-exported pub(crate) from mod.rs, so every external
crate::home::room_screen::* path is unchanged. Tests move with their
subjects (112 #[test] fns before and after, byte-identical bodies).
mod.rs: 17,790 -> 13,831 lines.
