+++ src/home/room_screen/mod.rs
@@ -58,6 +58,15 @@ use rangemap::RangeSet;
 
 use super::{ContextMenuOpenGesture, event_reaction_list::ReactionData, invite_modal::is_invite_modal_open, loading_pane::LoadingPaneRef, new_message_context_menu::{MessageAbilities, MessageDetails}, room_read_receipt::{self, populate_read_receipts, MAX_VISIBLE_AVATARS_IN_READ_RECEIPT}};
 
+mod bot_admin;
+mod bot_message;
+mod octos_actions;
+
+use bot_admin::*;
+use bot_message::*;
+use octos_actions::*;
+pub(crate) use bot_admin::is_known_or_likely_bot;
+
 /// The maximum number of timeline items to search through
 /// when looking for a particular event.
 ///
@@ -148,11 +157,6 @@ const BOT_BADGE_HORIZONTAL_PADDING: f64 = 6.0;
 const BOT_BADGE_BORDER_RADIUS: f64 = 3.0;
 const BOT_BADGE_TEXT_FONT_SIZE: f64 = 8.5;
 const BOT_BADGE_TEXT_TOP_DROP: f64 = -0.08;
-const MAX_OCTOS_ACTION_BUTTONS: usize = 6;
-const AGENTCHAT_APPROVAL_EVENT_KEY: &str = "com.agentchat.approval";
-const AGENTCHAT_APPROVAL_REQUEST_MSGTYPE: &str = "com.agentchat.approval.request.v1";
-const AGENTCHAT_APPROVAL_STATUS_MSGTYPE: &str = "com.agentchat.approval.status.v1";
-const AGENTCHAT_APPROVAL_VERDICT_MSGTYPE: &str = "com.agentchat.approval.verdict.v1";
 const MIN_SMALL_STATE_EVENTS_TO_COLLAPSE: usize = 2;
 
 const fn centered_top_margin(outer_top_margin: f64, outer_height: f64, inner_height: f64) -> f64 {
@@ -192,65 +196,6 @@ fn bot_badge_label_center_y() -> f64 {
     let bot_badge_label_top_margin = (BOT_BADGE_HEIGHT - BOT_BADGE_TEXT_FONT_SIZE) * 0.5
         + (BOT_BADGE_TEXT_FONT_SIZE * BOT_BADGE_TEXT_TOP_DROP);
     center_y(bot_badge_label_top_margin, BOT_BADGE_TEXT_FONT_SIZE)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct BotTimelineLayers {
-    status: Option<String>,
-    provider: Option<String>,
-    body: String,
-    footer: Option<String>,
-}
-
-impl BotTimelineLayers {
-    fn plain(body: &str) -> Self {
-        Self {
-            status: None,
-            provider: None,
-            body: body.to_string(),
-            footer: None,
-        }
-    }
-}
-
-/// How an agent-chat bridge tagged a relayed message: the bridge prefixes every
-/// agent message body with one of three emoji for the message `type`
-/// (`📋` request / `↩️` reply / `ℹ️` inform). It is the only *language-independent*
-/// structured signal on the wire, so it drives the badge instead of any prose.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AgentReplyKind {
-    Request,
-    Reply,
-    Inform,
-}
-
-impl AgentReplyKind {
-    /// Splits a leading type emoji off a bot body, if present.
-    /// Returns the kind plus the remaining text with the marker removed.
-    fn split_from_body(body: &str) -> (Option<Self>, &str) {
-        let trimmed = body.trim_start();
-        for (marker, kind) in [
-            ("📋", Self::Request),
-            ("↩️", Self::Reply),
-            // The variation-selector-free form of ℹ️ also occurs in the wild.
-            ("ℹ️", Self::Inform),
-            ("ℹ", Self::Inform),
-        ] {
-            if let Some(rest) = trimmed.strip_prefix(marker) {
-                return (Some(kind), rest.trim_start());
-            }
-        }
-        (None, body)
-    }
-
-    /// Short, translatable-later label shown in the badge next to the sender.
-    fn label(self) -> &'static str {
-        match self {
-            Self::Request => "request",
-            Self::Reply => "reply",
-            Self::Inform => "info",
-        }
-    }
 }
 
 /// How a message's delivery should be shown in the timeline.
@@ -317,1370 +262,6 @@ fn wedge_error_reason(error: &matrix_sdk::Error) -> String {
     }
 }
 
-/// The role an agent plays in an agent-chat workflow, derived from its Matrix
-/// localpart (`@ac_<team>_<role>:…`).
-///
-/// Every message in a workflow room otherwise carries the same generic `bot`
-/// badge, which says nothing about *who* is speaking — the coordinator handing
-/// off work, the implementer reporting a build, or a reviewer returning a
-/// verdict. The role is encoded in the account name, so reading it needs no
-/// prose parsing and is unaffected by the language the agent replies in.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum AgentRole {
-    Coordinator,
-    Implementer,
-    Reviewer,
-    FinalReviewer,
-}
-
-impl AgentRole {
-    /// Derives the role from a Matrix localpart, e.g. `ac_tyrese_reviewer`.
-    fn from_localpart(localpart: &str) -> Option<Self> {
-        let name = localpart.to_ascii_lowercase();
-        let matches = |suffix: &str| name == suffix || name.ends_with(&format!("_{suffix}"));
-        // `final_reviewer` also ends with `reviewer`, so it has to be tested
-        // first or every final reviewer would be labelled a plain reviewer.
-        if matches("final_reviewer") {
-            Some(Self::FinalReviewer)
-        } else if matches("coordinator") {
-            Some(Self::Coordinator)
-        } else if matches("implementer") {
-            Some(Self::Implementer)
-        } else if matches("reviewer") {
-            Some(Self::Reviewer)
-        } else {
-            None
-        }
-    }
-
-    /// Badge text. Kept lowercase to match the existing `bot` badge's voice.
-    fn label(self) -> &'static str {
-        match self {
-            Self::Coordinator => "coordinator",
-            Self::Implementer => "implementer",
-            Self::Reviewer => "reviewer",
-            Self::FinalReviewer => "final review",
-        }
-    }
-}
-
-/// Splits a trailing agent-chat permalink line (`🔗 https://…`) off a bot body.
-///
-/// The bridge appends this line to every relayed agent message. Left inline it
-/// is just a bare URL at the bottom of a wall of text — and it disappears
-/// entirely once the body is folded. Pulling it out lets the card pin it in a
-/// footer that stays reachable in both states.
-///
-/// Returns the body without that line, plus the URL.
-fn split_bot_permalink(body: &str) -> (String, Option<String>) {
-    let Some(last_line) = body.lines().next_back() else {
-        return (body.to_string(), None);
-    };
-    let trimmed = last_line.trim();
-    let Some(rest) = trimmed.strip_prefix('🔗') else {
-        return (body.to_string(), None);
-    };
-    let url = rest.trim();
-    if !(url.starts_with("http://") || url.starts_with("https://")) {
-        return (body.to_string(), None);
-    }
-    let kept: Vec<&str> = {
-        let mut lines: Vec<&str> = body.lines().collect();
-        lines.pop();
-        while lines.last().is_some_and(|line| line.trim().is_empty()) {
-            lines.pop();
-        }
-        lines
-    };
-    (kept.join("\n"), Some(url.to_string()))
-}
-
-/// Removes the bridge's trailing `<a …>🔗 View formatted</a>` anchor (and the
-/// `<br>`s leading up to it) from a formatted body, so the permalink is not
-/// rendered twice once it has been promoted to the card footer.
-fn strip_permalink_anchor_from_html(html: &str) -> String {
-    let Some(anchor_start) = html.rfind("<a ") else {
-        return html.to_string();
-    };
-    let tail = &html[anchor_start..];
-    if !tail.contains('🔗') || !tail.trim_end().ends_with("</a>") {
-        return html.to_string();
-    }
-    let mut head = html[..anchor_start].trim_end();
-    loop {
-        let trimmed = head
-            .strip_suffix("<br>")
-            .or_else(|| head.strip_suffix("<br/>"))
-            .or_else(|| head.strip_suffix("<br />"));
-        match trimmed {
-            Some(shorter) => head = shorter.trim_end(),
-            None => break,
-        }
-    }
-    head.to_string()
-}
-
-/// Bodies longer than this (in lines) are folded to a preview so one long
-/// message does not push every neighbouring one off-screen.
-const BOT_BODY_FOLD_LINE_THRESHOLD: usize = 8;
-/// How many lines of the body remain visible while folded.
-const BOT_BODY_FOLD_PREVIEW_LINES: usize = 3;
-/// Bodies longer than this (in characters) are folded even when they occupy few
-/// lines. A pasted JSON dump or a minified payload can be one enormous line: it
-/// wraps to fill the viewport while the line count says there is nothing to
-/// fold.
-const BOT_BODY_FOLD_CHAR_THRESHOLD: usize = 1200;
-/// How much of an over-long single line survives in the preview. Chosen to be a
-/// little more than the preview's line budget would show at a typical width, so
-/// the fold still reads as "the start of something", not a hard truncation.
-const BOT_BODY_FOLD_PREVIEW_CHARS: usize = 400;
-
-/// Folds `body` to a short preview when it is long enough to crowd the timeline.
-///
-/// Length is measured two ways, because a body can be oversized in either
-/// dimension: many lines, or few lines that are individually enormous. Returns
-/// `None` when the body is small on both counts — callers treat that as "not
-/// foldable", so the toggle stays hidden.
-fn fold_bot_body_preview(body: &str) -> Option<String> {
-    let lines: Vec<&str> = body.lines().collect();
-    let too_many_lines = lines.len() > BOT_BODY_FOLD_LINE_THRESHOLD;
-    let too_many_chars = body.chars().count() > BOT_BODY_FOLD_CHAR_THRESHOLD;
-    if !too_many_lines && !too_many_chars {
-        return None;
-    }
-
-    let preview: Vec<&str> = lines
-        .iter()
-        .copied()
-        .take(BOT_BODY_FOLD_PREVIEW_LINES)
-        .collect();
-    // A preview that ended up blank carries no information scent; showing the
-    // full body is better than an empty card with a "show more" affordance.
-    if preview.iter().all(|line| line.trim().is_empty()) {
-        return None;
-    }
-    let mut preview = preview.join("\n");
-
-    // Taking whole lines is not enough on its own: three lines of a minified
-    // payload can still be the whole screen. Cap the preview by characters too,
-    // on a char boundary so multi-byte text cannot panic the slice.
-    if preview.chars().count() > BOT_BODY_FOLD_PREVIEW_CHARS {
-        preview = preview.chars().take(BOT_BODY_FOLD_PREVIEW_CHARS).collect();
-        preview.push('…');
-    }
-
-    // Cutting mid-body can leave a ``` fence open, which would swallow the rest
-    // of the preview into an unterminated code block. Close it.
-    if preview.matches("```").count() % 2 == 1 {
-        preview.push_str("\n```");
-    }
-    Some(preview)
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct BotTimelineRenderState {
-    show_card: bool,
-    show_body_card: bool,
-    show_status_strip: bool,
-    show_metadata_footer: bool,
-    status: Option<String>,
-    provider: Option<String>,
-    body: String,
-    footer: Option<String>,
-    /// Message-type badge derived from the bridge's leading emoji marker.
-    kind: Option<AgentReplyKind>,
-    /// Set when `body` was long enough to fold; holds the folded preview text.
-    /// `None` means the body is short and the fold toggle must stay hidden.
-    folded_body: Option<String>,
-    /// The bridge's trailing permalink, promoted out of the body so the card can
-    /// pin it in a footer that stays visible while the body is folded.
-    permalink: Option<String>,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum OctosActionStyle {
-    Primary,
-    Secondary,
-    Danger,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct OctosActionButton {
-    id: String,
-    label: String,
-    style: OctosActionStyle,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ActionButtonRenderSlot {
-    id: String,
-    label: String,
-    style: OctosActionStyle,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct SelectedOctosActionState {
-    id: String,
-    label: String,
-    style: OctosActionStyle,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ApprovalCardRenderState {
-    title: String,
-    summary: String,
-    buttons_enabled: bool,
-    expired: bool,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ActionButtonRenderState {
-    show_container: bool,
-    show_button_row: bool,
-    approval_card: Option<ApprovalCardRenderState>,
-    buttons_enabled: bool,
-    visible_slots: Vec<ActionButtonRenderSlot>,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct ParsedOctosActionPayload {
-    approval_request: Option<OctosApprovalRequest>,
-    actions: Vec<OctosActionButton>,
-    malformed_approval_request: bool,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum OctosApprovalRiskLevel {
-    Normal,
-    Critical,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum OctosApprovalTimeoutBehavior {
-    Notify,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum ApprovalProtocol {
-    Octos,
-    AgentChat {
-        agent: String,
-        project: String,
-        project_room_id: String,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct OctosApprovalRequest {
-    protocol: ApprovalProtocol,
-    request_id: String,
-    tool_name: String,
-    tool_args_digest: String,
-    title: String,
-    summary: String,
-    risk_level: OctosApprovalRiskLevel,
-    authorized_approvers: Vec<String>,
-    expires_at: String,
-    on_timeout: OctosApprovalTimeoutBehavior,
-}
-
-fn parse_octos_action_style(style: Option<&str>) -> OctosActionStyle {
-    match style {
-        Some("primary") => OctosActionStyle::Primary,
-        Some("danger") => OctosActionStyle::Danger,
-        _ => OctosActionStyle::Secondary,
-    }
-}
-
-fn effective_octos_message_content(content: &serde_json::Value) -> &serde_json::Value {
-    content.get("m.new_content").unwrap_or(content)
-}
-
-fn latest_effective_event_content_json(
-    event_tl_item: &EventTimelineItem,
-) -> Option<serde_json::Value> {
-    event_tl_item.latest_edit_json()
-        .or_else(|| event_tl_item.original_json())
-        .and_then(|raw| raw.get_field::<serde_json::Value>("content").ok())
-        .flatten()
-        .map(|content| effective_octos_message_content(&content).clone())
-}
-
-fn original_event_content_json(
-    event_tl_item: &EventTimelineItem,
-) -> Option<serde_json::Value> {
-    event_tl_item.original_json()
-        .and_then(|raw| raw.get_field::<serde_json::Value>("content").ok())
-        .flatten()
-}
-
-fn event_raw_json_contains_any(
-    event_tl_item: &EventTimelineItem,
-    markers: &[&str],
-) -> bool {
-    event_tl_item
-        .latest_edit_json()
-        .into_iter()
-        .chain(event_tl_item.original_json())
-        .any(|raw| {
-            let json = raw.json().get();
-            markers.iter().any(|marker| json.contains(marker))
-        })
-}
-
-fn forwardable_room_message_content_from_json(
-    content: serde_json::Value,
-) -> Option<RoomMessageEventContent> {
-    let mut message = serde_json::from_value::<RoomMessageEventContent>(content).ok()?;
-    let is_forwardable = matches!(
-        &message.msgtype,
-        MessageType::Text(..) | MessageType::Notice(..) | MessageType::Emote(..)
-    );
-    message.relates_to = None;
-    message.tsp_signature = None;
-    is_forwardable.then_some(message)
-}
-
-fn parse_octos_approval_risk_level(value: Option<&str>) -> Option<OctosApprovalRiskLevel> {
-    match value {
-        Some("normal") => Some(OctosApprovalRiskLevel::Normal),
-        Some("critical") => Some(OctosApprovalRiskLevel::Critical),
-        _ => None,
-    }
-}
-
-fn parse_octos_approval_timeout_behavior(value: Option<&str>) -> Option<OctosApprovalTimeoutBehavior> {
-    match value {
-        Some("notify") => Some(OctosApprovalTimeoutBehavior::Notify),
-        _ => None,
-    }
-}
-
-fn parse_octos_approval_request_from_content(content: &serde_json::Value) -> Option<OctosApprovalRequest> {
-    let approval = content.get("org.octos.approval_request")?;
-    let request_id = approval.get("request_id")?.as_str()?.trim();
-    let tool_name = approval.get("tool_name")?.as_str()?.trim();
-    let tool_args_digest = approval.get("tool_args_digest")?.as_str()?.trim();
-    let title = approval.get("title")?.as_str()?.trim();
-    let summary = approval.get("summary")?.as_str()?.trim();
-    let risk_level = parse_octos_approval_risk_level(
-        approval.get("risk_level").and_then(|value| value.as_str()).map(str::trim),
-    )?;
-
-    let approvers = approval.get("authorized_approvers")?.as_array()?;
-    let authorized_approvers = approvers
-        .iter()
-        .filter_map(|value| value.as_str().map(str::trim))
-        .filter(|value| !value.is_empty())
-        .map(str::to_owned)
-        .collect::<Vec<_>>();
-    if authorized_approvers.is_empty() {
-        return None;
-    }
-
-    let expires_at = approval.get("expires_at")?.as_str()?.trim();
-    let on_timeout = parse_octos_approval_timeout_behavior(
-        approval.get("on_timeout").and_then(|value| value.as_str()).map(str::trim),
-    )?;
-
-    if request_id.is_empty()
-        || tool_name.is_empty()
-        || tool_args_digest.is_empty()
-        || title.is_empty()
-        || summary.is_empty()
-        || expires_at.is_empty()
-    {
-        return None;
-    }
-
-    Some(OctosApprovalRequest {
-        protocol: ApprovalProtocol::Octos,
-        request_id: request_id.to_owned(),
-        tool_name: tool_name.to_owned(),
-        tool_args_digest: tool_args_digest.to_owned(),
-        title: title.to_owned(),
-        summary: summary.to_owned(),
-        risk_level,
-        authorized_approvers,
-        expires_at: expires_at.to_owned(),
-        on_timeout,
-    })
-}
-
-fn is_lowercase_hex(value: &str, expected_len: usize) -> bool {
-    value.len() == expected_len
-        && value.bytes().all(|byte| byte.is_ascii_digit() || (b'a'..=b'f').contains(&byte))
-}
-
-fn parse_agentchat_approval_actions_from_detail(
-    approval: &serde_json::Value,
-) -> Vec<OctosActionButton> {
-    let Some(actions) = approval.get("actions").and_then(|value| value.as_array()) else {
-        return Vec::new();
-    };
-    if actions.len() != 2 {
-        return Vec::new();
-    }
-
-    let mut parsed = Vec::with_capacity(2);
-    for action in actions {
-        let Some(id) = action.get("id").and_then(|value| value.as_str()).map(str::trim) else {
-            return Vec::new();
-        };
-        let Some(label) = action.get("label").and_then(|value| value.as_str()).map(str::trim) else {
-            return Vec::new();
-        };
-        if label.is_empty() {
-            return Vec::new();
-        }
-        parsed.push(OctosActionButton {
-            id: id.to_owned(),
-            label: label.to_owned(),
-            style: parse_octos_action_style(action.get("style").and_then(|value| value.as_str())),
-        });
-    }
-
-    if parsed[0].id != "approve_once"
-        || parsed[0].style != OctosActionStyle::Primary
-        || parsed[1].id != "deny"
-        || parsed[1].style != OctosActionStyle::Danger
-    {
-        return Vec::new();
-    }
-    parsed
-}
-
-fn parse_agentchat_approval_request_from_content(
-    content: &serde_json::Value,
-) -> Option<OctosApprovalRequest> {
-    if content.get("msgtype").and_then(|value| value.as_str()) != Some(AGENTCHAT_APPROVAL_REQUEST_MSGTYPE) {
-        return None;
-    }
-    let approval = content.get(AGENTCHAT_APPROVAL_EVENT_KEY)?;
-    if approval.get("version").and_then(|value| value.as_u64()) != Some(1)
-        || approval.get("kind").and_then(|value| value.as_str()) != Some("request")
-        || parse_agentchat_approval_actions_from_detail(approval).len() != 2
-    {
-        return None;
-    }
-
-    let agent = approval.get("agent")?.as_str()?.trim();
-    let project = approval.get("project")?.as_str()?.trim();
-    let project_room_id = approval.get("project_room_id")?.as_str()?.trim();
-    let request_id = approval.get("request_id")?.as_str()?.trim();
-    let upstream_request_id = approval.get("upstream_request_id")?.as_str()?.trim();
-    let input_digest = approval.get("input_digest")?.as_str()?.trim();
-    let runtime = approval.get("runtime")?.as_str()?.trim();
-    let tool_name = approval.get("tool_name")?.as_str()?.trim();
-    let description = approval.get("description")?.as_str()?.trim();
-    let input_preview = approval.get("input_preview")?.as_str()?.trim();
-    let expires_at = approval.get("expires_at")?.as_u64()?;
-
-    let request_suffix = request_id.strip_prefix("approval_")?;
-    if agent.is_empty()
-        || project.is_empty()
-        || !project_room_id.starts_with('!')
-        || !project_room_id.contains(':')
-        || !is_lowercase_hex(request_suffix, 32)
-        || upstream_request_id.is_empty()
-        || !is_lowercase_hex(input_digest, 64)
-        || !matches!(runtime, "claude" | "codex")
-        || tool_name.is_empty()
-        || expires_at == 0
-    {
-        return None;
-    }
-
-    let summary = match (description.is_empty(), input_preview.is_empty()) {
-        (false, false) => format!("{description}\n{input_preview}"),
-        (false, true) => description.to_owned(),
-        (true, false) => input_preview.to_owned(),
-        (true, true) => project.to_owned(),
-    };
-
-    Some(OctosApprovalRequest {
-        protocol: ApprovalProtocol::AgentChat {
-            agent: agent.to_owned(),
-            project: project.to_owned(),
-            project_room_id: project_room_id.to_owned(),
-        },
-        request_id: request_id.to_owned(),
-        tool_name: tool_name.to_owned(),
-        tool_args_digest: input_digest.to_owned(),
-        title: tool_name.to_owned(),
-        summary,
-        risk_level: OctosApprovalRiskLevel::Normal,
-        authorized_approvers: Vec::new(),
-        expires_at: expires_at.to_string(),
-        on_timeout: OctosApprovalTimeoutBehavior::Notify,
-    })
-}
-
-fn agentchat_custom_message_body_from_content(content: &serde_json::Value) -> Option<&str> {
-    let msgtype = content.get("msgtype")?.as_str()?;
-    let expected_kind = match msgtype {
-        AGENTCHAT_APPROVAL_REQUEST_MSGTYPE => "request",
-        AGENTCHAT_APPROVAL_STATUS_MSGTYPE => "status",
-        AGENTCHAT_APPROVAL_VERDICT_MSGTYPE => "verdict",
-        _ => return None,
-    };
-    let approval = content.get(AGENTCHAT_APPROVAL_EVENT_KEY)?;
-    if approval.get("version").and_then(|value| value.as_u64()) != Some(1)
-        || approval.get("kind").and_then(|value| value.as_str()) != Some(expected_kind)
-    {
-        return None;
-    }
-    content.get("body")?.as_str()
-}
-
-fn parse_octos_actions_from_content(content: &serde_json::Value) -> Vec<OctosActionButton> {
-    let Some(actions) = effective_octos_message_content(content)
-        .get("org.octos.actions")
-        .and_then(|value| value.as_array())
-    else {
-        return Vec::new();
-    };
-
-    let mut parsed = Vec::new();
-    for (index, action) in actions.iter().enumerate() {
-        if parsed.len() >= MAX_OCTOS_ACTION_BUTTONS {
-            warning!(
-                "org.octos.actions: truncated {} extra buttons",
-                actions.len().saturating_sub(MAX_OCTOS_ACTION_BUTTONS)
-            );
-            break;
-        }
-
-        let Some(id) = action.get("id").and_then(|value| value.as_str()).map(str::trim) else {
-            warning!("org.octos.actions: skipping malformed entry at index {index}");
-            continue;
-        };
-        let Some(label) = action.get("label").and_then(|value| value.as_str()).map(str::trim) else {
-            warning!("org.octos.actions: skipping malformed entry at index {index}");
-            continue;
-        };
-        if id.is_empty() || label.is_empty() {
-            warning!("org.octos.actions: skipping malformed entry at index {index}");
-            continue;
-        }
-
-        parsed.push(OctosActionButton {
-            id: id.to_owned(),
-            label: label.to_owned(),
-            style: parse_octos_action_style(action.get("style").and_then(|value| value.as_str())),
-        });
-    }
-
-    parsed
-}
-
-fn parse_octos_approval_actions_from_content(content: &serde_json::Value) -> Vec<OctosActionButton> {
-    parse_octos_actions_from_content(content)
-        .into_iter()
-        .filter(|action| matches!(action.id.as_str(), "approve" | "deny"))
-        .collect()
-}
-
-fn parse_octos_action_payload_for_render(
-    content: Option<&serde_json::Value>,
-    original_content: Option<&serde_json::Value>,
-) -> ParsedOctosActionPayload {
-    let is_agentchat_request = original_content.is_some_and(|content| {
-        content.get("msgtype").and_then(|value| value.as_str()) == Some(AGENTCHAT_APPROVAL_REQUEST_MSGTYPE)
-    });
-    let has_octos_request = original_content
-        .is_some_and(|content| content.get("org.octos.approval_request").is_some());
-    let approval_request = if is_agentchat_request {
-        original_content.and_then(parse_agentchat_approval_request_from_content)
-    } else {
-        original_content.and_then(parse_octos_approval_request_from_content)
-    };
-    let malformed_approval_request = (is_agentchat_request || has_octos_request)
-        && approval_request.is_none();
-
-    let actions = if malformed_approval_request {
-        Vec::new()
-    } else if let Some(approval_request) = approval_request.as_ref() {
-        match approval_request.protocol {
-            ApprovalProtocol::Octos => original_content
-                .map(parse_octos_approval_actions_from_content)
-                .unwrap_or_default(),
-            ApprovalProtocol::AgentChat { .. } => original_content
-                .and_then(|content| content.get(AGENTCHAT_APPROVAL_EVENT_KEY))
-                .map(parse_agentchat_approval_actions_from_detail)
-                .unwrap_or_default(),
-        }
-    } else {
-        content
-            .map(parse_octos_actions_from_content)
-            .unwrap_or_default()
-    };
-
-    ParsedOctosActionPayload {
-        approval_request,
-        actions,
-        malformed_approval_request,
-    }
-}
-
-fn current_unix_time_millis() -> u64 {
-    SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .ok()
-        .and_then(|duration| u64::try_from(duration.as_millis()).ok())
-        .unwrap_or(u64::MAX)
-}
-
-fn approval_expiry_millis(protocol: &ApprovalProtocol, expires_at: &str) -> Option<u64> {
-    match protocol {
-        ApprovalProtocol::AgentChat { .. } => expires_at.parse().ok(),
-        ApprovalProtocol::Octos => chrono::DateTime::parse_from_rfc3339(expires_at)
-            .ok()
-            .and_then(|expires_at| u64::try_from(expires_at.timestamp_millis()).ok()),
-    }
-}
-
-fn approval_request_is_expired(
-    approval_request: &OctosApprovalRequest,
-    now_millis: u64,
-) -> bool {
-    approval_expiry_millis(&approval_request.protocol, &approval_request.expires_at)
-        .map(|expires_at| now_millis >= expires_at)
-        .unwrap_or(true)
-}
-
-fn compute_action_button_render_state_at(
-    actions: &[OctosActionButton],
-    approval_request: Option<&OctosApprovalRequest>,
-    current_user_id: Option<&UserId>,
-    now_millis: u64,
-) -> ActionButtonRenderState {
-    let approval_card = approval_request
-        .and_then(|approval_request| (!actions.is_empty()).then(|| {
-            let expired = approval_request_is_expired(approval_request, now_millis);
-            ApprovalCardRenderState {
-                title: approval_request.title.clone(),
-                summary: approval_request.summary.clone(),
-                buttons_enabled: !expired && local_user_can_approve(approval_request, current_user_id),
-                expired,
-            }
-        }));
-    let visible_slots = actions
-        .iter()
-        .take(MAX_OCTOS_ACTION_BUTTONS)
-        .map(|action| ActionButtonRenderSlot {
-            id: action.id.clone(),
-            label: action.label.clone(),
-            style: action.style,
-        })
-        .collect::<Vec<_>>();
-
-    let buttons_enabled = approval_card
-        .as_ref()
-        .map(|approval_card| approval_card.buttons_enabled)
-        .unwrap_or(true);
-    let show_button_row = !visible_slots.is_empty();
-
-    ActionButtonRenderState {
-        show_container: approval_card.is_some() || show_button_row,
-        show_button_row,
-        approval_card,
-        buttons_enabled,
-        visible_slots,
-    }
-}
-
-fn compute_action_button_render_state(
-    actions: &[OctosActionButton],
-    approval_request: Option<&OctosApprovalRequest>,
-    current_user_id: Option<&UserId>,
-) -> ActionButtonRenderState {
-    compute_action_button_render_state_at(
-        actions,
-        approval_request,
-        current_user_id,
-        current_unix_time_millis(),
-    )
-}
-
-fn action_button_render_slots_for_display(
-    render_state: &ActionButtonRenderState,
-    selected_action: Option<&SelectedOctosActionState>,
-) -> Vec<ActionButtonRenderSlot> {
-    if let Some(selected_action) = selected_action {
-        vec![ActionButtonRenderSlot {
-            id: selected_action.id.clone(),
-            label: format!("✓ {}", selected_action.label),
-            style: selected_action.style,
-        }]
-    } else {
-        render_state.visible_slots.clone()
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct OctosActionResponseRequest {
-    timeline_kind: TimelineKind,
-    content: serde_json::Value,
-    target_user_id: OwnedUserId,
-    explicit_room: bool,
-    source_event_id: OwnedEventId,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum OctosActionButtonRequest {
-    Generic {
-        action_id: String,
-        label: String,
-        style: OctosActionStyle,
-    },
-    Approval {
-        protocol: ApprovalProtocol,
-        request_id: String,
-        title: String,
-        decision: String,
-        label: String,
-        tool_args_digest: String,
-        expires_at: String,
-        style: OctosActionStyle,
-    },
-}
-
-impl OctosActionButtonRequest {
-    fn action_id(&self) -> &str {
-        match self {
-            Self::Generic { action_id, .. } => action_id,
-            Self::Approval { decision, .. } => decision,
-        }
-    }
-
-    fn label(&self) -> &str {
-        match self {
-            Self::Generic { label, .. } => label,
-            Self::Approval { label, .. } => label,
-        }
-    }
-
-    fn style(&self) -> OctosActionStyle {
-        match self {
-            Self::Generic { style, .. } | Self::Approval { style, .. } => *style,
-        }
-    }
-
-    fn expiry_millis(&self) -> Option<u64> {
-        match self {
-            Self::Generic { .. } => None,
-            Self::Approval { protocol, expires_at, .. } => {
-                approval_expiry_millis(protocol, expires_at)
-            }
-        }
-    }
-
-    fn is_expired(&self, now_millis: u64) -> bool {
-        self.expiry_millis()
-            .map(|expires_at| now_millis >= expires_at)
-            .unwrap_or(matches!(self, Self::Approval { .. }))
-    }
-}
-
-#[cfg(test)]
-fn next_approval_expiry_timeout<'a>(
-    requests: impl IntoIterator<Item = &'a OctosActionButtonRequest>,
-    now_millis: u64,
-) -> Option<Duration> {
-    earliest_approval_expiry_millis(requests)
-        .map(|expires_at| Duration::from_millis(
-            expires_at.saturating_sub(now_millis).max(1)
-        ))
-}
-
-fn earliest_approval_expiry_millis<'a>(
-    requests: impl IntoIterator<Item = &'a OctosActionButtonRequest>,
-) -> Option<u64> {
-    requests
-        .into_iter()
-        .filter_map(OctosActionButtonRequest::expiry_millis)
-        .min()
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct OctosActionButtonContext {
-    item_id: usize,
-    item_widget_uid: WidgetUid,
-    source_event_id: OwnedEventId,
-    original_sender: OwnedUserId,
-    request: OctosActionButtonRequest,
-}
-
-fn build_octos_approval_response_request(
-    timeline_kind: &TimelineKind,
-    title: &str,
-    request_id: &str,
-    decision: &str,
-    tool_args_digest: &str,
-    source_event_id: &EventId,
-    original_sender: &UserId,
-) -> OctosActionResponseRequest {
-    OctosActionResponseRequest {
-        timeline_kind: timeline_kind.clone(),
-        content: serde_json::json!({
-            "msgtype": "m.text",
-            "body": format!("[Approval: {decision}] {title}"),
-            "org.octos.approval_response": {
-                "request_id": request_id,
-                "decision": decision,
-                "source_event_id": source_event_id.as_str(),
-                "tool_args_digest": tool_args_digest,
-            },
-            "m.relates_to": {
-                "m.in_reply_to": {
-                    "event_id": source_event_id.as_str(),
-                }
-            }
-        }),
-        target_user_id: original_sender.to_owned(),
-        explicit_room: false,
-        source_event_id: source_event_id.to_owned(),
-    }
-}
-
-fn build_agentchat_approval_verdict_request(
-    timeline_kind: &TimelineKind,
-    label: &str,
-    request_id: &str,
-    action: &str,
-    input_digest: &str,
-    agent: &str,
-    project: &str,
-    project_room_id: &str,
-    source_event_id: &EventId,
-    original_sender: &UserId,
-) -> OctosActionResponseRequest {
-    OctosActionResponseRequest {
-        timeline_kind: timeline_kind.clone(),
-        content: serde_json::json!({
-            "msgtype": AGENTCHAT_APPROVAL_VERDICT_MSGTYPE,
-            "body": label,
-            "com.agentchat.approval": {
-                "version": 1,
-                "kind": "verdict",
-                "agent": agent,
-                "project": project,
-                "project_room_id": project_room_id,
-                "request_id": request_id,
-                "input_digest": input_digest,
-                "action": action,
-            },
-            "m.relates_to": {
-                "m.in_reply_to": {
-                    "event_id": source_event_id.as_str(),
-                }
-            }
-        }),
-        target_user_id: original_sender.to_owned(),
-        explicit_room: false,
-        source_event_id: source_event_id.to_owned(),
-    }
-}
-
-fn build_octos_action_response_request(
-    timeline_kind: &TimelineKind,
-    label: &str,
-    action_id: &str,
-    source_event_id: &EventId,
-    original_sender: &UserId,
-) -> OctosActionResponseRequest {
-    OctosActionResponseRequest {
-        timeline_kind: timeline_kind.clone(),
-        content: serde_json::json!({
-            "msgtype": "m.text",
-            "body": format!("[Action: {label}]"),
-            "org.octos.action_response": {
-                "action_id": action_id,
-                "source_event_id": source_event_id.as_str(),
-            },
-            "m.relates_to": {
-                "m.in_reply_to": {
-                    "event_id": source_event_id.as_str(),
-                }
-            }
-        }),
-        target_user_id: original_sender.to_owned(),
-        explicit_room: false,
-        source_event_id: source_event_id.to_owned(),
-    }
-}
-
-fn local_user_can_approve(
-    approval_request: &OctosApprovalRequest,
-    current_user_id: Option<&UserId>,
-) -> bool {
-    let Some(current_user_id) = current_user_id else {
-        return false;
-    };
-
-    match &approval_request.protocol {
-        ApprovalProtocol::Octos => approval_request.authorized_approvers
-            .iter()
-            .any(|approver| approver == current_user_id.as_str()),
-        // The dedicated encrypted approval room only contains the owner and
-        // managed service accounts. This enables the UI affordance, but is not
-        // an authorization decision: agent-chat still validates event.sender.
-        ApprovalProtocol::AgentChat { .. } => true,
-    }
-}
-
-fn mark_action_buttons_disabled(
-    disabled_source_event_ids: &mut HashSet<OwnedEventId>,
-    source_event_id: &OwnedEventId,
-) {
-    disabled_source_event_ids.insert(source_event_id.clone());
-}
-
-fn mark_selected_octos_action(
-    selected_actions: &mut HashMap<OwnedEventId, SelectedOctosActionState>,
-    source_event_id: &OwnedEventId,
-    action_id: &str,
-    label: &str,
-    style: OctosActionStyle,
-) {
-    selected_actions.insert(source_event_id.clone(), SelectedOctosActionState {
-        id: action_id.to_owned(),
-        label: label.to_owned(),
-        style,
-    });
-}
-
-fn clear_selected_octos_action(
-    selected_actions: &mut HashMap<OwnedEventId, SelectedOctosActionState>,
-    source_event_id: &EventId,
-) {
-    selected_actions.remove(source_event_id);
-}
-
-fn clear_action_buttons_disabled(
-    disabled_source_event_ids: &mut HashSet<OwnedEventId>,
-    source_event_id: &EventId,
-) {
-    disabled_source_event_ids.remove(source_event_id);
-}
-
-fn are_action_buttons_disabled(
-    disabled_source_event_ids: &HashSet<OwnedEventId>,
-    source_event_id: &EventId,
-) -> bool {
-    disabled_source_event_ids.contains(source_event_id)
-}
-
-fn octos_action_button_paths(index: usize) -> (&'static [LiveId], &'static [LiveId], &'static [LiveId], &'static [LiveId]) {
-    match index {
-        0 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_0)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_0), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_0), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_0), live_id!(danger_button)],
-        ),
-        1 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_1)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_1), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_1), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_1), live_id!(danger_button)],
-        ),
-        2 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_2)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_2), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_2), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_2), live_id!(danger_button)],
-        ),
-        3 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_3)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_3), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_3), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_3), live_id!(danger_button)],
-        ),
-        4 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_4)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_4), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_4), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_4), live_id!(danger_button)],
-        ),
-        _ => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_5)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_5), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_5), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(action_button_row), live_id!(action_button_slot_5), live_id!(danger_button)],
-        ),
-    }
-}
-
-fn approval_action_button_paths(index: usize) -> (&'static [LiveId], &'static [LiveId], &'static [LiveId], &'static [LiveId]) {
-    match index {
-        0 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_0)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_0), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_0), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_0), live_id!(danger_button)],
-        ),
-        1 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_1)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_1), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_1), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_1), live_id!(danger_button)],
-        ),
-        2 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_2)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_2), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_2), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_2), live_id!(danger_button)],
-        ),
-        3 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_3)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_3), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_3), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_3), live_id!(danger_button)],
-        ),
-        4 => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_4)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_4), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_4), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_4), live_id!(danger_button)],
-        ),
-        _ => (
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_5)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_5), live_id!(primary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_5), live_id!(secondary_button)],
-            &[live_id!(content), live_id!(action_buttons), live_id!(approval_request_view), live_id!(approval_action_button_row), live_id!(approval_button_slot_5), live_id!(danger_button)],
-        ),
-    }
-}
-
-fn is_bot_provider_line(line: &str) -> bool {
-    let trimmed = line.trim();
-    trimmed.starts_with("via ") && trimmed.contains('(') && trimmed.ends_with(')')
-}
-
-fn strip_streaming_cursor_suffix(line: &str) -> &str {
-    line
-        .trim_end()
-        .strip_suffix('\u{25CF}')
-        .map(str::trim_end)
-        .unwrap_or_else(|| line.trim_end())
-}
-
-fn is_bot_footer_line(line: &str) -> bool {
-    let trimmed = strip_streaming_cursor_suffix(line);
-    trimmed.starts_with('_')
-        && trimmed.ends_with('_')
-        && trimmed.contains("·")
-        && trimmed.contains(" in")
-        && trimmed.contains(" out")
-}
-
-fn looks_like_metrics_line(line: &str) -> bool {
-    let trimmed = strip_streaming_cursor_suffix(line).trim();
-    !trimmed.is_empty()
-        && trimmed.chars().count() <= 40
-        && trimmed.chars().any(|ch| ch.is_ascii_digit())
-        && (trimmed.contains('s') || trimmed.contains(" in") || trimmed.contains(" out"))
-}
-
-fn looks_like_status_line(line: &str) -> bool {
-    let trimmed = line.trim();
-    !trimmed.is_empty()
-        && !trimmed.starts_with("via ")
-        && !trimmed.starts_with('_')
-        && trimmed.chars().count() <= 32
-        && !trimmed.contains("  ")
-}
-
-fn trim_structured_body_lines(lines: &[&str]) -> String {
-    let mut start = 0;
-    let mut end = lines.len();
-
-    while start < end && lines[start].trim().is_empty() {
-        start += 1;
-    }
-    while end > start && lines[end - 1].trim().is_empty() {
-        end -= 1;
-    }
-
-    lines[start..end].join("\n")
-}
-
-fn is_viable_bot_body(body: &str) -> bool {
-    let trimmed = body.trim();
-    !trimmed.is_empty()
-        && trimmed.chars().any(|c| c.is_alphanumeric())
-}
-
-fn parse_bot_timeline_layers(raw_body: &str, is_bot_sender: bool) -> BotTimelineLayers {
-    if !is_bot_sender || raw_body.trim().is_empty() {
-        return BotTimelineLayers::plain(raw_body);
-    }
-
-    let lines: Vec<&str> = raw_body.lines().collect();
-    if lines.is_empty() {
-        return BotTimelineLayers::plain(raw_body);
-    }
-
-    let (status, provider, mut content_start) =
-        if lines.len() >= 2 && looks_like_status_line(lines[0]) && is_bot_provider_line(lines[1]) {
-            (
-                Some(lines[0].trim().to_string()),
-                Some(lines[1].trim().to_string()),
-                2usize,
-            )
-        } else if is_bot_provider_line(lines[0]) {
-            (None, Some(lines[0].trim().to_string()), 1usize)
-        } else {
-            (None, None, 0usize)
-        };
-
-    while content_start < lines.len() && lines[content_start].trim().is_empty() {
-        content_start += 1;
-    }
-
-    let mut footer = None;
-    let mut content_end = lines.len();
-    let last_non_empty = lines.iter().rposition(|line| !line.trim().is_empty());
-
-    if let Some(last_idx) = last_non_empty {
-        if is_bot_footer_line(lines[last_idx]) {
-            footer = Some(strip_streaming_cursor_suffix(lines[last_idx]).trim().to_string());
-            content_end = last_idx;
-            while content_end > content_start && lines[content_end - 1].trim().is_empty() {
-                content_end -= 1;
-            }
-        }
-    }
-
-    if content_start >= content_end {
-        return if status.is_some() || provider.is_some() || footer.is_some() {
-            BotTimelineLayers {
-                status,
-                provider,
-                body: String::new(),
-                footer,
-            }
-        } else {
-            BotTimelineLayers::plain(raw_body)
-        };
-    }
-
-    let content_lines = &lines[content_start..content_end];
-    let mut body = trim_structured_body_lines(content_lines);
-    if footer.is_none() && content_lines.len() == 1 && looks_like_metrics_line(content_lines[0]) {
-        footer = Some(strip_streaming_cursor_suffix(content_lines[0]).trim().to_string());
-        body.clear();
-    }
-    if !is_viable_bot_body(&body) {
-        return if status.is_some() || provider.is_some() || footer.is_some() {
-            BotTimelineLayers {
-                status,
-                provider,
-                body,
-                footer,
-            }
-        } else {
-            BotTimelineLayers::plain(raw_body)
-        };
-    }
-
-    BotTimelineLayers {
-        status,
-        provider,
-        body,
-        footer,
-    }
-}
-
-fn compute_bot_timeline_render_state(raw_body: &str, is_bot_sender: bool) -> BotTimelineRenderState {
-    let layers = parse_bot_timeline_layers(raw_body, is_bot_sender);
-    let show_card = is_bot_sender;
-
-    // Strip the bridge's leading type emoji off the body and keep it as a badge
-    // instead: it is redundant as prose but valuable as a typed affordance.
-    let (kind, body, permalink) = if show_card {
-        let (kind, rest) = AgentReplyKind::split_from_body(&layers.body);
-        // Promote the trailing permalink out of the body before measuring it for
-        // folding, so the preview spends its lines on content rather than a URL.
-        let (body, permalink) = split_bot_permalink(rest);
-        (kind, body, permalink)
-    } else {
-        (None, layers.body, None)
-    };
-
-    let show_body_card = show_card && !body.trim().is_empty();
-    let folded_body = if show_body_card {
-        fold_bot_body_preview(&body)
-    } else {
-        None
-    };
-
-    BotTimelineRenderState {
-        show_card,
-        show_body_card,
-        show_status_strip: show_card && layers.status.is_some(),
-        show_metadata_footer: show_card && (layers.provider.is_some() || layers.footer.is_some()),
-        status: layers.status,
-        provider: layers.provider,
-        body,
-        footer: layers.footer,
-        kind,
-        folded_body,
-        permalink,
-    }
-}
-
-/// The text that should land on the clipboard when copying a message body.
-///
-/// Bot-sent messages embed status / provider / `_metadata_` scaffolding lines
-/// in their raw body; the bubble strips them at render time via
-/// `compute_bot_timeline_render_state`, so copying must strip them the same
-/// way. Human messages are copied verbatim.
-fn clipboard_text_for_message_body(body: String, sender_is_bot: bool) -> String {
-    if sender_is_bot {
-        compute_bot_timeline_render_state(&body, true).body
-    } else {
-        body
-    }
-}
-
-fn display_bot_footer_text(footer: &str) -> &str {
-    strip_streaming_cursor_suffix(footer)
-        .strip_prefix('_')
-        .and_then(|trimmed| trimmed.strip_suffix('_'))
-        .unwrap_or(footer)
-}
-
-fn has_rich_markdown_syntax(text: &str) -> bool {
-    let trimmed = text.trim();
-    !trimmed.is_empty()
-        && (
-            trimmed.contains("```")
-            || trimmed.starts_with("## ")
-            || trimmed.starts_with("### ")
-            || trimmed.contains("\n## ")
-            || trimmed.contains("\n### ")
-            || trimmed.starts_with("|")
-            || trimmed.contains("\n|")
-            || trimmed.starts_with("- ")
-            || trimmed.contains("\n- ")
-            || trimmed.starts_with("* ")
-            || trimmed.contains("\n* ")
-            || trimmed.contains("**")
-            || trimmed.contains("`")
-        )
-}
-
-fn should_render_streaming_full_snapshot(
-    body: &str,
-    formatted_body: Option<&FormattedBody>,
-    is_bot_sender: bool,
-) -> bool {
-    is_bot_sender
-        && (
-            formatted_body.is_some_and(|formatted| formatted.format == MessageFormat::Html)
-            || has_rich_markdown_syntax(body)
-        )
-}
-
-fn select_bot_timeline_body_formatted_body(
-    render_state: &BotTimelineRenderState,
-    formatted_body: Option<&FormattedBody>,
-) -> Option<FormattedBody> {
-    if render_state.status.is_none()
-        && render_state.provider.is_none()
-        && render_state.footer.is_none()
-    {
-        return formatted_body
-            .cloned()
-            .or_else(|| has_rich_markdown_syntax(&render_state.body)
-                .then(|| FormattedBody::markdown(&render_state.body))
-                .flatten());
-    }
-
-    FormattedBody::markdown(&render_state.body)
-}
-
-fn should_render_bot_timeline_body_with_markdown_widget(
-    render_state: &BotTimelineRenderState,
-) -> bool {
-    render_state.show_body_card
-        && render_state.body.contains("```")
-}
-
-fn contains_cjk(text: &str) -> bool {
-    text.chars().any(|ch|
-        matches!(ch as u32,
-            0x3400..=0x4DBF
-            | 0x4E00..=0x9FFF
-            | 0xF900..=0xFAFF
-            | 0x20000..=0x2A6DF
-            | 0x2A700..=0x2B73F
-            | 0x2B740..=0x2B81F
-            | 0x2B820..=0x2CEAF
-            | 0x2CEB0..=0x2EBEF
-            | 0x3000..=0x303F
-            | 0x3040..=0x30FF
-            | 0x31F0..=0x31FF
-            | 0xAC00..=0xD7AF
-        )
-    )
-}
-
-fn fenced_code_blocks_contain_cjk(text: &str) -> bool {
-    let mut in_fence = false;
-    let mut fence_has_cjk = false;
-
-    for line in text.lines() {
-        if line.trim_start().starts_with("```") {
-            if in_fence && fence_has_cjk {
-                return true;
-            }
-            in_fence = !in_fence;
-            fence_has_cjk = false;
-            continue;
-        }
-
-        if in_fence && contains_cjk(line) {
-            fence_has_cjk = true;
-        }
-    }
-
-    in_fence && fence_has_cjk
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum BotTimelineCodeBlockMode {
-    None,
-    Highlighted,
-    Plain,
-}
-
-fn bot_timeline_code_block_mode(render_state: &BotTimelineRenderState) -> BotTimelineCodeBlockMode {
-    if !should_render_bot_timeline_body_with_markdown_widget(render_state) {
-        return BotTimelineCodeBlockMode::None;
-    }
-
-    if fenced_code_blocks_contain_cjk(&render_state.body) {
-        BotTimelineCodeBlockMode::Plain
-    } else {
-        BotTimelineCodeBlockMode::Highlighted
-    }
-}
-
-fn streaming_update_requires_content_invalidation(
-    state: &StreamingAnimState,
-    new_text: &str,
-    is_live: bool,
-    render_full_target: bool,
-) -> bool {
-    state.target_text != new_text
-        || state.is_live != is_live
-        || state.render_full_target != render_full_target
-}
-
 thread_local! {
     static ROOM_INFO_ACTION_MODAL_OPEN: Cell<bool> = const { Cell::new(false) };
 }
@@ -1709,33 +290,6 @@ fn item_event_id(item: &Arc<TimelineItem>) -> Option<&EventId> {
     event.event_id()
 }
 
-/// Check if an event carries the MSC4357 `org.matrix.msc4357.live` field,
-/// indicating that the message content is still being streamed.
-///
-/// For edit events (`m.replace`), the live field lives inside `m.new_content`
-/// rather than at the top level of `content`, so we check both locations.
-fn content_has_msc4357_live_marker(content: &serde_json::Value) -> bool {
-    let effective = content.get("m.new_content").unwrap_or(content);
-    match effective.get("org.matrix.msc4357.live") {
-        Some(serde_json::Value::Bool(value)) => *value,
-        Some(_) => true,
-        None => false,
-    }
-}
-
-fn is_msc4357_live(event_tl_item: &EventTimelineItem) -> bool {
-    let message_is_edited = event_tl_item
-        .content()
-        .as_message()
-        .is_some_and(|message| message.is_edited());
-    event_tl_item.latest_edit_json()
-        .or_else(|| (!message_is_edited).then(|| event_tl_item.original_json()).flatten())
-        .and_then(|raw| raw.get_field::<serde_json::Value>("content").ok())
-        .flatten()
-        .map(|content| content_has_msc4357_live_marker(&content))
-        .unwrap_or(false)
-}
-
 fn compute_translation_lang_popup_abs_pos(button_rect: Rect, container_rect: Rect) -> DVec2 {
     let min_x = container_rect.pos.x + TRANSLATION_LANG_POPUP_MARGIN;
     let max_x = (container_rect.pos.x + container_rect.size.x - TRANSLATION_LANG_POPUP_WIDTH - TRANSLATION_LANG_POPUP_MARGIN)
@@ -1757,554 +311,6 @@ fn compute_translation_lang_popup_abs_pos(button_rect: Rect, container_rect: Rec
     };
 
     dvec2(popup_x, popup_y)
-}
-
-fn streaming_scan_range(
-    clear_cache: bool,
-    changed_indices: &Range<usize>,
-    _old_len: usize,
-    new_len: usize,
-) -> Range<usize> {
-    if clear_cache {
-        0..new_len
-    } else {
-        let start = changed_indices.start.min(new_len);
-        let end = changed_indices.end.min(new_len);
-        start..end
-    }
-}
-
-fn refresh_stream_indices<'a, I>(
-    event_ids: I,
-    streaming_messages: &mut HashMap<OwnedEventId, super::streaming_animation::StreamingAnimState>,
-)
-where
-    I: IntoIterator<Item = Option<&'a EventId>>,
-{
-    for state in streaming_messages.values_mut() {
-        state.timeline_index = None;
-    }
-
-    for (idx, event_id) in event_ids.into_iter().enumerate() {
-        let Some(event_id) = event_id else {
-            continue;
-        };
-        if let Some(state) = streaming_messages.get_mut(event_id) {
-            state.timeline_index = Some(idx);
-        }
-    }
-}
-
-fn any_timeline_indices_visible<I, F>(
-    indices: I,
-    is_visible: F,
-) -> bool
-where
-    I: IntoIterator<Item = Option<usize>>,
-    F: FnMut(usize) -> bool,
-{
-    indices.into_iter().flatten().any(is_visible)
-}
-
-fn streaming_candidates_from_items<'a>(
-    items: &'a Vector<Arc<TimelineItem>>,
-) -> impl Iterator<Item = (OwnedEventId, String, bool)> + 'a {
-    items.iter().filter_map(|item| {
-        let TimelineItemKind::Event(event) = item.kind() else {
-            return None;
-        };
-        let event_id = event.event_id()?.to_owned();
-        let text = RoomScreen::extract_message_text(item)?;
-        Some((event_id, text, is_msc4357_live(event)))
-    })
-}
-
-fn rebuild_streaming_messages_for_full_snapshot<I>(
-    items: I,
-    previous_streaming_messages: Option<&HashMap<OwnedEventId, super::streaming_animation::StreamingAnimState>>,
-) -> (HashMap<OwnedEventId, super::streaming_animation::StreamingAnimState>, bool)
-where
-    I: IntoIterator<Item = (OwnedEventId, String, bool)>,
-{
-    use crate::home::streaming_animation::StreamingAnimState;
-
-    let mut rebuilt = HashMap::new();
-    let mut should_schedule_frame = false;
-
-    for (event_id, new_text, live) in items {
-        if !live {
-            continue;
-        }
-
-        // Only restore animations that were already tracked before the
-        // snapshot reset.  Never create brand-new animations here — during
-        // initial/reconnect loads the SDK may not have aggregated edits yet,
-        // so completed messages can still appear as `live`.  Genuinely new
-        // streams will be picked up on the next live sync update.
-        if let Some(previous_state) = previous_streaming_messages
-            .and_then(|states| states.get(&event_id))
-        {
-            let state = StreamingAnimState::restore(previous_state, &new_text, true);
-            should_schedule_frame |= state.needs_frame();
-            rebuilt.insert(event_id, state);
-        }
-    }
-
-    (rebuilt, should_schedule_frame)
-}
-
-fn next_stream_timeout<'a>(
-    states: impl IntoIterator<Item = &'a super::streaming_animation::StreamingAnimState>,
-) -> Option<Duration> {
-    states
-        .into_iter()
-        .map(|state| state.timeout_after().saturating_sub(state.last_update_time.elapsed()))
-        .min()
-}
-
-fn escape_slash_command_arg(value: &str) -> String {
-    value.trim().replace('\\', "\\\\").replace('"', "\\\"")
-}
-
-fn format_create_bot_command(
-    username: &str,
-    display_name: &str,
-    system_prompt: Option<&str>,
-) -> String {
-    let mut command = format!("/createbot {} {}", username.trim(), display_name.trim());
-    if let Some(system_prompt) = system_prompt
-        .map(str::trim)
-        .filter(|value| !value.is_empty())
-    {
-        command.push_str(" --prompt \"");
-        command.push_str(&escape_slash_command_arg(system_prompt));
-        command.push('"');
-    }
-    command
-}
-
-fn format_delete_bot_command(matrix_user_id: &UserId) -> String {
-    format!("/deletebot {matrix_user_id}")
-}
-
-fn resolve_delete_bot_user_id(
-    user_id_or_localpart: &str,
-    current_user_id: Option<&UserId>,
-    app_language: AppLanguage,
-) -> Result<OwnedUserId, String> {
-    let raw = user_id_or_localpart.trim();
-    if raw.is_empty() {
-        return Err(tr_key(app_language, "room_screen.bot.delete.error.empty_user_id").into());
-    }
-
-    if raw.starts_with('@') || raw.contains(':') {
-        let full_user_id = if raw.starts_with('@') {
-            raw.to_string()
-        } else {
-            format!("@{raw}")
-        };
-        return UserId::parse(&full_user_id)
-            .map(|user_id| user_id.to_owned())
-            .map_err(|_| tr_fmt(app_language, "room_screen.bot.delete.error.invalid_user_id", &[
-                ("full_user_id", full_user_id.as_str()),
-            ]));
-    }
-
-    let Some(current_user_id) = current_user_id else {
-        return Err(
-            tr_key(app_language, "room_screen.bot.delete.error.current_user_unavailable").into(),
-        );
-    };
-
-    let full_user_id = format!("@{raw}:{}", current_user_id.server_name());
-    UserId::parse(&full_user_id)
-        .map(|user_id| user_id.to_owned())
-        .map_err(|_| tr_fmt(app_language, "room_screen.bot.delete.error.invalid_user_id", &[
-            ("full_user_id", full_user_id.as_str()),
-        ]))
-}
-
-fn detected_bot_binding_for_members(
-    app_state: &AppState,
-    room_id: &OwnedRoomId,
-    members: &[RoomMember],
-) -> Option<OwnedUserId> {
-    if app_state.bot_settings.is_room_bound(room_id) {
-        return None;
-    }
-
-    let own_user_id = current_user_id();
-    let is_non_self = |room_member: &&RoomMember| {
-        own_user_id
-            .as_deref()
-            .is_none_or(|own_user_id| room_member.user_id() != own_user_id)
-    };
-
-    if let Ok(configured_bot_user_id) = app_state
-        .bot_settings
-        .resolved_bot_user_id(current_user_id().as_deref())
-    {
-        if members
-            .iter()
-            .filter(is_non_self)
-            .any(|room_member| room_member.user_id().as_str() == configured_bot_user_id.as_str())
-        {
-            return Some(configured_bot_user_id);
-        }
-    }
-
-    let known_bot_user_ids = timeline_known_bot_user_ids(app_state);
-    if let Some(bot_member) = members
-        .iter()
-        .filter(is_non_self)
-        .filter(|room_member|
-            known_bot_user_ids
-                .iter()
-                .any(|known_bot_user_id| known_bot_user_id.as_str() == room_member.user_id().as_str())
-        )
-        .min_by(|lhs, rhs| lhs.user_id().as_str().cmp(rhs.user_id().as_str()))
-    {
-        return Some(bot_member.user_id().to_owned());
-    }
-
-    let mut non_self_members = members.iter().filter(is_non_self);
-    if let Some(dm_counterparty) = non_self_members.next()
-        && non_self_members.next().is_none()
-    {
-        let localpart = dm_counterparty.user_id().localpart().to_ascii_lowercase();
-        let localpart_likely_bot = localpart == "bot"
-            || localpart == "botfather"
-            || localpart.starts_with("bot_")
-            || localpart.starts_with("bot-")
-            || localpart.starts_with("bot.");
-        let display_name_likely_bot = dm_counterparty
-            .display_name()
-            .is_some_and(|display_name| display_name.to_ascii_lowercase().contains("bot"));
-        if localpart_likely_bot || display_name_likely_bot {
-            return Some(dm_counterparty.user_id().to_owned());
-        }
-    }
-
-    members
-        .iter()
-        .filter(is_non_self)
-        .filter(|room_member| room_member.user_id().localpart().eq_ignore_ascii_case("botfather"))
-        .min_by(|lhs, rhs| lhs.user_id().as_str().cmp(rhs.user_id().as_str()))
-        .map(|room_member| room_member.user_id().to_owned())
-}
-
-fn is_likely_bot_user_id(
-    user_id: &UserId,
-    resolved_parent_bot_user_id: Option<&UserId>,
-) -> bool {
-    if resolved_parent_bot_user_id.is_some_and(|parent| parent == user_id) {
-        return true;
-    }
-
-    let localpart = user_id.localpart().to_ascii_lowercase();
-    localpart == "bot"
-        || localpart == "botfather"
-        || localpart.starts_with("bot_")
-        || localpart.starts_with("bot-")
-        || localpart.starts_with("bot.")
-        || localpart.ends_with("_bot")
-        || (localpart.ends_with("bot") && localpart.len() > 3)
-        || is_agent_chat_puppet_localpart(&localpart)
-}
-
-/// Agent-chat puppets an agent per Matrix account named
-/// `<MATRIX_AGENT_PREFIX><team>_<role>` — the prefix defaults to `ac_` and the
-/// roles come from the shared issue-workflow skill. None of those names contain
-/// "bot", so without this they render as human messages and miss the bot card
-/// (and its type badge / fold affordance) entirely.
-fn is_agent_chat_puppet_localpart(localpart: &str) -> bool {
-    const AGENT_CHAT_PREFIX: &str = "ac_";
-    const WORKFLOW_ROLE_SUFFIXES: &[&str] = &[
-        "_coordinator",
-        "_implementer",
-        "_reviewer",
-        "_final_reviewer",
-    ];
-
-    localpart.starts_with(AGENT_CHAT_PREFIX)
-        || WORKFLOW_ROLE_SUFFIXES
-            .iter()
-            .any(|suffix| localpart.ends_with(suffix))
-}
-
-pub(crate) fn is_known_or_likely_bot(
-    user_id: &UserId,
-    resolved_parent_bot_user_id: Option<&UserId>,
-    known_bot_user_ids: &[OwnedUserId],
-) -> bool {
-    known_bot_user_ids
-        .iter()
-        .any(|known_bot_user_id| known_bot_user_id.as_str() == user_id.as_str())
-        || resolved_parent_bot_user_id.is_some_and(|parent| parent == user_id)
-        || is_likely_bot_user_id(user_id, resolved_parent_bot_user_id)
-}
-
-fn is_timeline_sender_bot(
-    user_id: &UserId,
-    resolved_parent_bot_user_id: Option<&UserId>,
-    room_bot_user_ids: &[OwnedUserId],
-    known_bot_user_ids: &[OwnedUserId],
-) -> bool {
-    room_bot_user_ids
-        .iter()
-        .any(|room_bot_user_id| room_bot_user_id.as_str() == user_id.as_str())
-        || is_known_or_likely_bot(
-            user_id,
-            resolved_parent_bot_user_id,
-            known_bot_user_ids,
-        )
-}
-
-fn collect_room_bot_user_ids(
-    room_members: &[RoomMember],
-    resolved_parent_bot_user_id: Option<&UserId>,
-    known_bot_user_ids: &[OwnedUserId],
-    persisted_room_bot_user_ids: &[OwnedUserId],
-) -> Vec<OwnedUserId> {
-    let own_user_id = current_user_id();
-    let mut room_bot_user_ids = Vec::<OwnedUserId>::new();
-
-    for persisted_room_bot_user_id in persisted_room_bot_user_ids {
-        if room_bot_user_ids
-            .iter()
-            .all(|existing_user_id| existing_user_id.as_str() != persisted_room_bot_user_id.as_str())
-        {
-            room_bot_user_ids.push(persisted_room_bot_user_id.clone());
-        }
-    }
-
-    for room_member in room_members.iter().filter(|room_member|
-        own_user_id
-            .as_deref()
-            .is_none_or(|own_user_id| room_member.user_id() != own_user_id)
-    ) {
-        if is_known_or_likely_bot(
-            room_member.user_id(),
-            resolved_parent_bot_user_id,
-            known_bot_user_ids,
-        ) || is_likely_bot_member(room_member, resolved_parent_bot_user_id)
-        {
-            let user_id = room_member.user_id().to_owned();
-            if room_bot_user_ids
-                .iter()
-                .all(|existing_user_id| existing_user_id.as_str() != user_id.as_str())
-            {
-                room_bot_user_ids.push(user_id);
-            }
-        }
-    }
-
-    room_bot_user_ids.sort_by(|lhs, rhs| lhs.as_str().cmp(rhs.as_str()));
-    room_bot_user_ids
-}
-
-/// Returns the set of MXIDs the timeline should treat as bots: the union of the
-/// app-service known-bot list (only when app-service is enabled) and every agent
-/// registered in the global [`AgentRegistry`] (always, independent of app-service).
-fn timeline_known_bot_user_ids(app_state: &AppState) -> Vec<OwnedUserId> {
-    let mut bot_user_ids = if app_state.bot_settings.enabled {
-        app_state.bot_settings.known_bot_user_ids()
-    } else {
-        Vec::new()
-    };
-    for agent_user_id in app_state.agent_registry.agent_user_ids() {
-        if bot_user_ids
-            .iter()
-            .all(|existing| existing.as_str() != agent_user_id.as_str())
-        {
-            bot_user_ids.push(agent_user_id);
-        }
-    }
-    bot_user_ids
-}
-
-#[derive(Clone, Default)]
-struct TimelineBotContext {
-    app_service_enabled: bool,
-    app_service_room_bound: bool,
-    has_persisted_management_binding: bool,
-    bound_bot_user_id: Option<OwnedUserId>,
-    resolved_parent_bot_user_id: Option<OwnedUserId>,
-    persisted_bound_bot_user_ids: Vec<OwnedUserId>,
-    room_bot_user_ids: Vec<OwnedUserId>,
-    known_bot_user_ids: Vec<OwnedUserId>,
-}
-
-struct CachedTimelineBotContext {
-    room_id: OwnedRoomId,
-    room_members: Option<Arc<Vec<RoomMember>>>,
-    app_service_enabled: bool,
-    room_is_bound: bool,
-    persisted_bound_bot_user_id: Option<OwnedUserId>,
-    persisted_bound_bot_user_ids: Vec<OwnedUserId>,
-    resolved_parent_bot_user_id: Option<OwnedUserId>,
-    known_bot_user_ids: Vec<OwnedUserId>,
-    value: TimelineBotContext,
-}
-impl CachedTimelineBotContext {
-    fn has_same_members(&self, room_members: Option<&Arc<Vec<RoomMember>>>) -> bool {
-        match (self.room_members.as_ref(), room_members) {
-            (Some(cached), Some(current)) => Arc::ptr_eq(cached, current),
-            (None, None) => true,
-            _ => false,
-        }
-    }
-
-    fn matches(
-        &self,
-        room_id: &OwnedRoomId,
-        room_members: Option<&Arc<Vec<RoomMember>>>,
-        app_service_enabled: bool,
-        room_is_bound: bool,
-        persisted_bound_bot_user_id: Option<&OwnedUserId>,
-        persisted_bound_bot_user_ids: &[OwnedUserId],
-        resolved_parent_bot_user_id: Option<&OwnedUserId>,
-        known_bot_user_ids: &[OwnedUserId],
-    ) -> bool {
-        self.room_id == *room_id
-            && self.has_same_members(room_members)
-            && self.app_service_enabled == app_service_enabled
-            && self.room_is_bound == room_is_bound
-            && self.persisted_bound_bot_user_id.as_ref() == persisted_bound_bot_user_id
-            && self.persisted_bound_bot_user_ids.as_slice() == persisted_bound_bot_user_ids
-            && self.resolved_parent_bot_user_id.as_ref() == resolved_parent_bot_user_id
-            && self.known_bot_user_ids.as_slice() == known_bot_user_ids
-    }
-}
-
-fn compute_timeline_bot_context(
-    app_state: Option<&AppState>,
-    room_id: &OwnedRoomId,
-    room_members: Option<&Arc<Vec<RoomMember>>>,
-) -> (Option<OwnedUserId>, Vec<OwnedUserId>, Vec<OwnedUserId>) {
-    app_state
-        .map(|app_state| {
-            let app_service_enabled = app_state.bot_settings.enabled;
-            let persisted_room_bot_user_ids = if app_service_enabled {
-                app_state.bot_settings.bound_bot_user_ids(room_id)
-            } else {
-                Vec::new()
-            };
-            let resolved_parent_bot_user_id = if app_service_enabled {
-                app_state
-                    .bot_settings
-                    .resolved_bot_user_id(current_user_id().as_deref())
-                    .ok()
-            } else {
-                None
-            };
-            // Union of the (app-service-gated) known-bot list and the global
-            // AgentRegistry, so registry agents are recognized even when the
-            // app-service integration is disabled.
-            let known_bot_user_ids = timeline_known_bot_user_ids(app_state);
-            let room_bot_user_ids = room_members
-                .map(|members|
-                    collect_room_bot_user_ids(
-                        members.as_ref(),
-                        resolved_parent_bot_user_id.as_deref(),
-                        &known_bot_user_ids,
-                        &persisted_room_bot_user_ids,
-                    )
-                )
-                .unwrap_or(persisted_room_bot_user_ids);
-            (
-                resolved_parent_bot_user_id,
-                room_bot_user_ids,
-                known_bot_user_ids,
-            )
-        })
-        .unwrap_or((None, Vec::new(), Vec::new()))
-}
-
-fn is_likely_bot_member(
-    room_member: &RoomMember,
-    resolved_parent_bot_user_id: Option<&UserId>,
-) -> bool {
-    if is_likely_bot_user_id(room_member.user_id(), resolved_parent_bot_user_id) {
-        return true;
-    }
-
-    room_member.display_name().is_some_and(|display_name| {
-        let display_name = display_name.trim().to_ascii_lowercase();
-        display_name == "bot"
-            || display_name == "botfather"
-            || display_name.starts_with("bot ")
-            || display_name.ends_with(" bot")
-            || display_name.contains(" bot ")
-    })
-}
-
-fn extract_bot_user_ids_from_listbots_reply(
-    text: &str,
-    default_server_name: Option<&OwnedServerName>,
-) -> Vec<OwnedUserId> {
-    let mut bot_user_ids = Vec::<OwnedUserId>::new();
-
-    let mut push_bot = |bot_user_id: OwnedUserId| {
-        if !bot_user_ids
-            .iter()
-            .any(|existing_bot_user_id| existing_bot_user_id.as_str() == bot_user_id.as_str())
-        {
-            bot_user_ids.push(bot_user_id);
-        }
-    };
-
-    for token in text.split(|ch: char|
-        !(ch.is_ascii_alphanumeric() || matches!(ch, '@' | ':' | '_' | '-' | '.'))
-    ) {
-        let token = token.trim();
-        if token.is_empty() {
-            continue;
-        }
-
-        if token.starts_with('@') && token.contains(':') {
-            if let Ok(bot_user_id) = UserId::parse(token).map(|user_id| user_id.to_owned()) {
-                if BotSettingsState::is_valid_known_bot_user_id(bot_user_id.as_ref()) {
-                    push_bot(bot_user_id);
-                }
-            }
-            continue;
-        }
-
-        if token.contains(':') && !token.starts_with('@') {
-            let full_user_id = format!("@{token}");
-            if let Ok(bot_user_id) = UserId::parse(&full_user_id).map(|user_id| user_id.to_owned()) {
-                if BotSettingsState::is_valid_known_bot_user_id(bot_user_id.as_ref()) {
-                    push_bot(bot_user_id);
-                }
-            }
-            continue;
-        }
-
-        let localpart_lc = token.to_ascii_lowercase();
-        let is_likely_bot_localpart = (
-                localpart_lc == "bot"
-                || localpart_lc.starts_with("bot_")
-                || localpart_lc.starts_with("bot-")
-                || localpart_lc.starts_with("bot.")
-            )
-            && localpart_lc != "bots"
-            && localpart_lc != "botfather"
-            && token
-                .chars()
-                .all(|ch| ch.is_ascii_alphanumeric() || ch == '_' || ch == '-' || ch == '.');
-        if !is_likely_bot_localpart {
-            continue;
-        }
-
-        let Some(default_server_name) = default_server_name else { continue };
-        let full_user_id = format!("@{token}:{default_server_name}");
-        if let Ok(bot_user_id) = UserId::parse(&full_user_id).map(|user_id| user_id.to_owned()) {
-            push_bot(bot_user_id);
-        }
-    }
-
-    bot_user_ids
 }
 
 script_mod! {
@@ -13692,38 +11698,6 @@ fn populate_text_message_content(
     }
 }
 
-/// Labels the bot badge with the sender's workflow role, when it has one.
-///
-/// A workflow agent gets its role (`coordinator`, `reviewer`, …) in the accent
-/// style, so the participants of a run stand out from incidental bots, which
-/// keep the generic `bot` label in the quieter neutral style.
-///
-/// Both the text and both colors are always written, never only on the branch
-/// that needs them: these item widgets are recycled by the PortalList, so a
-/// value left unset would keep whatever the previously drawn message put there.
-fn populate_bot_badge_identity(cx: &mut Cx, item: &WidgetRef, sender_localpart: &str) {
-    let role = AgentRole::from_localpart(sender_localpart);
-    let mut badge = item.view(cx, ids!(content.username_view.bot_badge));
-    let mut label = item.label(cx, ids!(content.username_view.bot_badge.bot_badge_label));
-
-    label.set_text(cx, role.map_or("bot", AgentRole::label));
-    if role.is_some() {
-        script_apply_eval!(cx, badge, {
-            draw_bg +: { color: (mod.widgets.RBX_ACCENT_SOFT) }
-        });
-        script_apply_eval!(cx, label, {
-            draw_text +: { color: (mod.widgets.RBX_ACCENT) }
-        });
-    } else {
-        script_apply_eval!(cx, badge, {
-            draw_bg +: { color: (mod.widgets.RBX_NEUTRAL_BG) }
-        });
-        script_apply_eval!(cx, label, {
-            draw_text +: { color: (mod.widgets.RBX_NEUTRAL_FG) }
-        });
-    }
-}
-
 /// Renders a message's delivery state: a pill in the meta band, plus the reason
 /// and the two recovery actions when the send is parked for good.
 ///
@@ -13777,432 +11751,6 @@ fn populate_send_state(cx: &mut Cx, item: &WidgetRef, event_tl_item: &EventTimel
                 .set_text(cx, reason);
         }
         _ => failure.set_visible(cx, false),
-    }
-}
-
-/// Puts a `Message` item back into its plain (non-bot-card) state.
-///
-/// `populate_bot_text_message_content` is the only code that flips these two
-/// visibilities, and it only runs on the text/notice path. The PortalList hands
-/// the *same* `Message` widget to the server-notice and unsupported/redacted
-/// paths as well, so a slot that previously held a bot reply would keep drawing
-/// that reply's card while the new content sat hidden underneath it. Every
-/// branch that renders into `content.message` must call this first.
-fn reset_bot_message_card(cx: &mut Cx, item: &WidgetRef) {
-    item.view(cx, ids!(content.bot_message_card)).set_visible(cx, false);
-    item.html_or_plaintext(cx, ids!(content.message)).set_visible(cx, true);
-}
-
-fn populate_bot_text_message_content(
-    cx: &mut Cx,
-    item: &WidgetRef,
-    app_language: AppLanguage,
-    body: &str,
-    formatted_body: Option<&FormattedBody>,
-    room_mention_room_id: Option<&OwnedRoomId>,
-    link_preview_ref: Option<&mut LinkPreviewRef>,
-    media_cache: Option<&mut MediaCache>,
-    link_preview_cache: Option<&mut LinkPreviewCache>,
-    is_bot_sender: bool,
-    bot_body_expanded: bool,
-    // True while the body is still being streamed in; suppresses folding.
-    is_streaming: bool,
-) -> (bool, Option<String>) {
-    let render_state = compute_bot_timeline_render_state(body, is_bot_sender);
-    let bot_card_view = item.view(cx, ids!(content.bot_message_card));
-    let message_view = item.html_or_plaintext(cx, ids!(content.message));
-
-    bot_card_view.set_visible(cx, render_state.show_card);
-    message_view.set_visible(cx, !render_state.show_card);
-
-    if !render_state.show_card {
-        // Clear the meta-band badge before returning: this item widget is
-        // recycled by the PortalList, so a stale badge from a previously drawn
-        // bot message would otherwise linger on a plain/human message.
-        item.view(cx, ids!(content.message_action_bar.kind_badge))
-            .set_visible(cx, false);
-
-        // A plain message can crowd the timeline just as badly as an agent
-        // report — a pasted log, or one enormous line of JSON. Fold it the same
-        // way, with the toggle in the meta band.
-        let folded = fold_bot_body_preview(body);
-        let is_folded = folded.is_some() && !bot_body_expanded;
-        let toggle = item.button(cx, ids!(content.message_action_bar.plain_fold_toggle));
-        toggle.set_visible(cx, folded.is_some());
-        if folded.is_some() {
-            toggle.set_text(cx, if is_folded { "Show more" } else { "Show less" });
-        }
-
-        let drawn = if is_folded {
-            // The preview is a truncated slice of the plain body, so the
-            // message's `formatted_body` no longer describes it, and link
-            // previews belong to content the user has not asked to see.
-            populate_text_message_content(
-                cx,
-                &message_view,
-                app_language,
-                folded.as_deref().unwrap_or(body),
-                None,
-                room_mention_room_id,
-                None,
-                None,
-                None,
-            )
-        } else {
-            populate_text_message_content(
-                cx,
-                &message_view,
-                app_language,
-                body,
-                formatted_body,
-                room_mention_room_id,
-                link_preview_ref,
-                media_cache,
-                link_preview_cache,
-            )
-        };
-        return (drawn, None);
-    }
-    // Bot cards own their fold toggle in the card footer.
-    item.button(cx, ids!(content.message_action_bar.plain_fold_toggle))
-        .set_visible(cx, false);
-
-    let status_strip = item.view(cx, ids!(content.bot_message_card.bot_status_strip));
-    status_strip.set_visible(cx, render_state.show_status_strip);
-    if let Some(status) = render_state.status.as_ref() {
-        item.label(cx, ids!(content.bot_message_card.bot_status_strip.bot_status_label))
-            .set_text(cx, status);
-    }
-
-    // Type badge: shown only when the bridge marked the message type.
-    // The type badge lives in the meta band (next to the copy icon), so it adds
-    // no vertical space of its own.
-    let kind_badge = item.view(cx, ids!(content.message_action_bar.kind_badge));
-    kind_badge.set_visible(cx, render_state.kind.is_some());
-    if let Some(kind) = render_state.kind {
-        item.label(cx, ids!(content.message_action_bar.kind_badge.kind_badge_label))
-            .set_text(cx, kind.label());
-    }
-
-    // The provider/footer metadata is rendered by the meta band below the card
-    // (content.message_action_bar.metadata_label), joined into a single line.
-    let band_metadata = if render_state.show_metadata_footer {
-        match (render_state.provider.as_ref(), render_state.footer.as_ref()) {
-            (Some(provider), Some(footer)) =>
-                Some(format!("{provider} · {}", display_bot_footer_text(footer))),
-            (Some(provider), None) => Some(provider.clone()),
-            (None, Some(footer)) => Some(display_bot_footer_text(footer).to_string()),
-            (None, None) => None,
-        }
-    } else {
-        None
-    };
-
-    let body_card = item.view(cx, ids!(content.bot_message_card.bot_body_card));
-    body_card.set_visible(cx, render_state.show_body_card);
-    let body_widget = item.html_or_plaintext(cx, ids!(content.bot_message_card.bot_body_card.bot_card_body));
-    let mut markdown_widget = item.markdown(cx, ids!(content.bot_message_card.bot_body_card.bot_card_markdown));
-    let mut markdown_plain_widget = item.markdown(cx, ids!(content.bot_message_card.bot_body_card.bot_card_markdown_plain));
-    // Fold state: a long body renders its preview until the user expands it.
-    // `folded_body` is `None` for short bodies, so the toggle stays hidden and
-    // the full text renders exactly as before.
-    //
-    // A streaming reply is never folded. Its body grows on every frame, so the
-    // moment it crosses the threshold the text the user is watching appear would
-    // collapse to three lines and take the typewriter cursor with it — the reply
-    // reads as having stalled. Folding applies once the message has settled.
-    let foldable = render_state.folded_body.is_some() && !is_streaming;
-    let is_folded = foldable && !bot_body_expanded;
-    let fold_toggle = item.button(
-        cx,
-        ids!(content.bot_message_card.bot_body_card.bot_card_footer_row.bot_body_fold_toggle),
-    );
-    fold_toggle.set_visible(cx, foldable);
-    if foldable {
-        fold_toggle.set_text(cx, if is_folded { "Show more" } else { "Show less" });
-    }
-
-    // Permalink pinned in the card footer — visible whether or not the body is
-    // folded, and clickable via the room screen's existing `HtmlLinkAction`
-    // handler.
-    let permalink_link = item.link_label(
-        cx,
-        ids!(content.bot_message_card.bot_body_card.bot_card_footer_row.bot_permalink_link),
-    );
-    permalink_link.set_visible(cx, render_state.permalink.is_some());
-    if let Some(url) = render_state.permalink.as_ref() {
-        permalink_link.set_text(cx, "View formatted");
-        if let Some(mut inner) = permalink_link.borrow_mut() {
-            inner.url = url.clone();
-        }
-    }
-
-    // Folding swaps the body for its preview *inside the render state* rather
-    // than switching rendering modes. Every downstream decision (code-block
-    // mode, formatted-body selection, which widget is visible) is then derived
-    // from the text actually on screen — one path, one visible widget. Choosing
-    // a different mode for the folded case instead would leave the preview in
-    // the plaintext widget and the full text in the markdown widget, and both
-    // would render (a duplicated body).
-    let render_state = if is_folded {
-        let mut folded = render_state.clone();
-        folded.body = folded.folded_body.clone().unwrap_or(folded.body);
-        folded
-    } else {
-        render_state
-    };
-    // A folded preview never carries the message's `formatted_body`: that HTML
-    // describes the full text, not this truncated slice. When the permalink has
-    // been promoted to the footer, drop its anchor from the HTML too so it does
-    // not render a second time inside the body.
-    let formatted_body_owned;
-    let formatted_body = if is_folded {
-        None
-    } else if render_state.permalink.is_some() {
-        match formatted_body {
-            Some(fb) => {
-                let mut stripped = fb.clone();
-                stripped.body = strip_permalink_anchor_from_html(&fb.body);
-                formatted_body_owned = stripped;
-                Some(&formatted_body_owned)
-            }
-            None => None,
-        }
-    } else {
-        formatted_body
-    };
-
-    let code_block_mode = bot_timeline_code_block_mode(&render_state);
-    body_widget.set_visible(cx, code_block_mode == BotTimelineCodeBlockMode::None);
-    markdown_widget.set_visible(cx, code_block_mode == BotTimelineCodeBlockMode::Highlighted);
-    markdown_plain_widget.set_visible(cx, code_block_mode == BotTimelineCodeBlockMode::Plain);
-    // Hiding the inactive renderer is not enough: a Markdown widget that still
-    // holds text keeps drawing its own DrawList after `set_visible(false)`, so
-    // toggling the fold on a body with a ``` block stacked the folded preview on
-    // top of the leftover full text. Worse, that leftover is invisible for
-    // hit-testing, so its links stopped responding to clicks. Clearing the text
-    // of whichever renderer is not in use makes it draw nothing at all.
-    if code_block_mode != BotTimelineCodeBlockMode::Highlighted {
-        markdown_widget.set_text(cx, "");
-    }
-    if code_block_mode != BotTimelineCodeBlockMode::Plain {
-        markdown_plain_widget.set_text(cx, "");
-    }
-    if code_block_mode != BotTimelineCodeBlockMode::None {
-        body_widget.show_plaintext(cx, "");
-    }
-
-    let drawn = if render_state.show_body_card {
-        if code_block_mode != BotTimelineCodeBlockMode::None {
-            match code_block_mode {
-                BotTimelineCodeBlockMode::Highlighted => markdown_widget.set_text(cx, &render_state.body),
-                BotTimelineCodeBlockMode::Plain => markdown_plain_widget.set_text(cx, &render_state.body),
-                BotTimelineCodeBlockMode::None => { }
-            }
-
-            if let (Some(link_preview_ref), Some(media_cache), Some(link_preview_cache)) =
-                (link_preview_ref, media_cache, link_preview_cache)
-            {
-                let mut links = Vec::new();
-                let _ = utils::linkify_get_urls(&render_state.body, false, Some(&mut links));
-                link_preview_ref.populate_below_message(
-                    cx,
-                    &links,
-                    media_cache,
-                    link_preview_cache,
-                    &|cx, text_or_image_ref, image_info_source, original_source, body, media_cache| {
-                        populate_image_message_content(
-                            cx,
-                            text_or_image_ref,
-                            None,
-                            app_language,
-                            image_info_source,
-                            original_source,
-                            body,
-                            media_cache,
-                        )
-                    },
-                )
-            } else {
-                true
-            }
-        } else {
-            let formatted_body_for_card =
-                select_bot_timeline_body_formatted_body(&render_state, formatted_body);
-            populate_text_message_content(
-                cx,
-                &body_widget,
-                app_language,
-                &render_state.body,
-                formatted_body_for_card.as_ref(),
-                room_mention_room_id,
-                link_preview_ref,
-                media_cache,
-                link_preview_cache,
-            )
-        }
-    } else {
-        true
-    };
-    (drawn, band_metadata)
-}
-
-fn populate_octos_action_buttons(
-    cx: &mut Cx,
-    app_language: AppLanguage,
-    item: &WidgetRef,
-    item_id: usize,
-    content: Option<&serde_json::Value>,
-    original_content: Option<&serde_json::Value>,
-    source_event_id: Option<&OwnedEventId>,
-    original_sender: &UserId,
-    action_button_contexts: &mut HashMap<WidgetUid, OctosActionButtonContext>,
-    disabled_source_event_ids: &HashSet<OwnedEventId>,
-    selected_actions: &HashMap<OwnedEventId, SelectedOctosActionState>,
-) {
-    let container = item.view(cx, ids!(content.action_buttons));
-    if content.is_none() && original_content.is_none() {
-        action_button_contexts.retain(|_, context| context.item_id != item_id);
-        container.set_visible(cx, false);
-        return;
-    }
-    let Some(source_event_id) = source_event_id else {
-        action_button_contexts.retain(|_, context| context.item_id != item_id);
-        container.set_visible(cx, false);
-        return;
-    };
-
-    let parsed_payload = parse_octos_action_payload_for_render(content, original_content);
-
-    if parsed_payload.malformed_approval_request {
-        warning!("approval request: skipping malformed structured payload");
-    }
-
-    let render_state = compute_action_button_render_state(
-        &parsed_payload.actions,
-        parsed_payload.approval_request.as_ref(),
-        current_user_id().as_deref(),
-    );
-    let is_disabled = are_action_buttons_disabled(disabled_source_event_ids, source_event_id.as_ref())
-        || !render_state.buttons_enabled;
-    let selected_action = selected_actions.get(source_event_id);
-    let visible_slots = action_button_render_slots_for_display(&render_state, selected_action);
-    let is_approval = render_state.approval_card.is_some();
-
-    container.set_visible(cx, render_state.show_container);
-    if !render_state.show_container {
-        action_button_contexts.retain(|_, context| context.item_id != item_id);
-        return;
-    }
-
-    let approval_request_view = item.view(cx, ids!(content.action_buttons.approval_request_view));
-    let button_row = item.view(cx, ids!(content.action_buttons.action_button_row));
-    let approval_button_row = item.view(cx, ids!(content.action_buttons.approval_request_view.approval_action_button_row));
-    button_row.set_visible(cx, !is_approval && render_state.show_button_row && !visible_slots.is_empty());
-    approval_button_row.set_visible(cx, is_approval && render_state.show_button_row && !visible_slots.is_empty());
-    approval_request_view.set_visible(cx, is_approval);
-    if let Some(approval_card) = render_state.approval_card.as_ref() {
-        item.label(cx, ids!(content.action_buttons.approval_request_view.approval_header.approval_title_label))
-            .set_text(cx, &approval_card.title);
-        item.label(cx, ids!(content.action_buttons.approval_request_view.approval_summary_label))
-            .set_text(cx, &approval_card.summary);
-        item.label(cx, ids!(content.action_buttons.approval_request_view.approval_header.pending_badge.pending_label))
-            .set_text(cx, tr_key(
-                app_language,
-                if approval_card.expired {
-                    "room_screen.approval.expired"
-                } else {
-                    "room_screen.approval.pending"
-                },
-            ));
-    }
-
-    for index in 0..MAX_OCTOS_ACTION_BUTTONS {
-        let (slot_path, primary_path, secondary_path, danger_path) = octos_action_button_paths(index);
-        item.view(cx, slot_path).set_visible(cx, false);
-
-        let primary_button = item.button(cx, primary_path);
-        action_button_contexts.remove(&primary_button.widget_uid());
-        primary_button.set_visible(cx, false);
-        primary_button.set_enabled(cx, !is_disabled);
-
-        let secondary_button = item.button(cx, secondary_path);
-        action_button_contexts.remove(&secondary_button.widget_uid());
-        secondary_button.set_visible(cx, false);
-        secondary_button.set_enabled(cx, !is_disabled);
-
-        let danger_button = item.button(cx, danger_path);
-        action_button_contexts.remove(&danger_button.widget_uid());
-        danger_button.set_visible(cx, false);
-        danger_button.set_enabled(cx, !is_disabled);
-
-        let (approval_slot_path, approval_primary_path, approval_secondary_path, approval_danger_path) =
-            approval_action_button_paths(index);
-        item.view(cx, approval_slot_path).set_visible(cx, false);
-
-        let approval_primary_button = item.button(cx, approval_primary_path);
-        action_button_contexts.remove(&approval_primary_button.widget_uid());
-        approval_primary_button.set_visible(cx, false);
-        approval_primary_button.set_enabled(cx, !is_disabled);
-
-        let approval_secondary_button = item.button(cx, approval_secondary_path);
-        action_button_contexts.remove(&approval_secondary_button.widget_uid());
-        approval_secondary_button.set_visible(cx, false);
-        approval_secondary_button.set_enabled(cx, !is_disabled);
-
-        let approval_danger_button = item.button(cx, approval_danger_path);
-        action_button_contexts.remove(&approval_danger_button.widget_uid());
-        approval_danger_button.set_visible(cx, false);
-        approval_danger_button.set_enabled(cx, !is_disabled);
-
-        let Some(render_slot) = visible_slots.get(index) else { continue };
-        let active_button = if is_approval {
-            item.view(cx, approval_slot_path).set_visible(cx, true);
-            match render_slot.style {
-                OctosActionStyle::Primary => approval_primary_button,
-                OctosActionStyle::Secondary => approval_secondary_button,
-                OctosActionStyle::Danger => approval_danger_button,
-            }
-        } else {
-            item.view(cx, slot_path).set_visible(cx, true);
-            match render_slot.style {
-                OctosActionStyle::Primary => primary_button,
-                OctosActionStyle::Secondary => secondary_button,
-                OctosActionStyle::Danger => danger_button,
-            }
-        };
-        active_button.set_visible(cx, true);
-        active_button.set_enabled(cx, !is_disabled);
-        active_button.set_text(cx, &render_slot.label);
-
-        if !is_disabled {
-            let request = if let Some(approval_request) = parsed_payload.approval_request.as_ref() {
-                OctosActionButtonRequest::Approval {
-                    protocol: approval_request.protocol.clone(),
-                    request_id: approval_request.request_id.clone(),
-                    title: approval_request.title.clone(),
-                    decision: render_slot.id.clone(),
-                    label: render_slot.label.clone(),
-                    tool_args_digest: approval_request.tool_args_digest.clone(),
-                    expires_at: approval_request.expires_at.clone(),
-                    style: render_slot.style,
-                }
-            } else {
-                OctosActionButtonRequest::Generic {
-                    action_id: render_slot.id.clone(),
-                    label: render_slot.label.clone(),
-                    style: render_slot.style,
-                }
-            };
-
-            action_button_contexts.insert(active_button.widget_uid(), OctosActionButtonContext {
-                item_id,
-                item_widget_uid: item.widget_uid(),
-                source_event_id: source_event_id.clone(),
-                original_sender: original_sender.to_owned(),
-                request,
-            });
-        }
     }
 }
 
@@ -15854,12 +13402,6 @@ pub fn clear_timeline_states(_cx: &mut Cx) {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::home::streaming_animation::StreamingAnimState;
-    use std::time::{Duration, Instant};
-
-    fn make_state(text: &str) -> StreamingAnimState {
-        StreamingAnimState::new(text, true)
-    }
 
     #[test]
     fn adjacent_timeline_snapshots_coalesce_without_crossing_control_updates() {
@@ -15978,64 +13520,6 @@ mod tests {
     }
 
     #[test]
-    fn test_forward_menu() {
-        let content = serde_json::json!({
-            "msgtype": "m.text",
-            "body": "hello"
-        });
-
-        let message = forwardable_room_message_content_from_json(content).unwrap();
-
-        assert!(matches!(message.msgtype, MessageType::Text(..)));
-    }
-
-    #[test]
-    fn test_forward_menu_hidden_non_message() {
-        let content = serde_json::json!({
-            "msgtype": "m.image",
-            "body": "photo.jpg",
-            "url": "mxc://example.org/media"
-        });
-
-        assert!(forwardable_room_message_content_from_json(content).is_none());
-    }
-
-    #[test]
-    fn test_forward_uses_latest_effective_content() {
-        let content = serde_json::json!({
-            "msgtype": "m.text",
-            "body": "original",
-            "m.new_content": {
-                "msgtype": "m.text",
-                "body": "edited"
-            }
-        });
-        let effective_content = effective_octos_message_content(&content).clone();
-        let message = forwardable_room_message_content_from_json(effective_content).unwrap();
-
-        assert!(matches!(
-            message.msgtype,
-            MessageType::Text(TextMessageEventContent { body, .. }) if body == "edited"
-        ));
-    }
-
-    #[test]
-    fn test_forward_does_not_send_reply_metadata() {
-        let content = serde_json::json!({
-            "msgtype": "m.text",
-            "body": "reply text",
-            "m.relates_to": {
-                "m.in_reply_to": {
-                    "event_id": "$source:example.org"
-                }
-            }
-        });
-        let message = forwardable_room_message_content_from_json(content).unwrap();
-
-        assert!(message.relates_to.is_none());
-    }
-
-    #[test]
     fn test_notice_offset_actions() {
         assert_eq!(tl_idx_from_item_id(0, true), None);
         assert_eq!(tl_idx_from_item_id(1, true), Some(0));
@@ -16044,100 +13528,6 @@ mod tests {
         assert_eq!(item_id_from_tl_idx(0, true), 1);
         assert_eq!(item_id_from_tl_idx(6, true), 7);
         assert_eq!(item_id_from_tl_idx(6, false), 6);
-    }
-
-    #[test]
-    fn test_streaming_scan_range() {
-        // Incremental: clamp sentinel to new_len
-        assert_eq!(streaming_scan_range(false, &(5..usize::MAX), 8, 9), 5..9);
-        // Append: new item at end is scanned
-        assert_eq!(streaming_scan_range(false, &(8..9), 8, 9), 8..9);
-        // No changes: empty range
-        assert_eq!(streaming_scan_range(false, &(8..8), 8, 8), 8..8);
-        // Clear cache: full scan
-        assert_eq!(streaming_scan_range(true, &(5..usize::MAX), 8, 9), 0..9);
-    }
-
-    #[test]
-    fn test_refresh_stream_indices() {
-        let event_id_a: OwnedEventId = "$event-a:example.com".try_into().unwrap();
-        let event_id_b: OwnedEventId = "$event-b:example.com".try_into().unwrap();
-        let missing_event_id: OwnedEventId = "$missing:example.com".try_into().unwrap();
-
-        let mut streaming_messages = HashMap::new();
-        streaming_messages.insert(event_id_a.clone(), make_state("alpha"));
-        streaming_messages.insert(missing_event_id.clone(), make_state("missing"));
-
-        let event_ids = vec![None, Some(event_id_a.as_ref()), Some(event_id_b.as_ref())];
-        refresh_stream_indices(event_ids, &mut streaming_messages);
-
-        assert_eq!(streaming_messages[&event_id_a].timeline_index, Some(1));
-        assert_eq!(streaming_messages[&missing_event_id].timeline_index, None);
-    }
-
-    #[test]
-    fn test_timeout_picks_earliest() {
-        let mut live = make_state("alpha");
-        live.last_update_time = Instant::now() - Duration::from_secs(40);
-        let mut finished = make_state("beta");
-        finished.is_live = false;
-        finished.last_update_time = Instant::now() - Duration::from_secs(29);
-
-        let timeout = next_stream_timeout([&live, &finished]).unwrap();
-
-        assert!(timeout <= Duration::from_secs(1));
-    }
-
-    #[test]
-    fn test_full_snapshot_rebuild_drops_finished_cached_streams() {
-        let event_id: OwnedEventId = "$event-live:example.com".try_into().unwrap();
-        let mut previous = HashMap::new();
-        let mut previous_state = make_state("hello live");
-        previous_state.advance_displayed(4);
-        previous.insert(event_id.clone(), previous_state);
-
-        let (rebuilt, should_schedule_frame) = rebuild_streaming_messages_for_full_snapshot(
-            [(event_id, String::from("hello final"), false)],
-            Some(&previous),
-        );
-
-        assert!(rebuilt.is_empty());
-        assert!(!should_schedule_frame);
-    }
-
-    #[test]
-    fn test_full_snapshot_rebuild_restores_live_cached_streams() {
-        let event_id: OwnedEventId = "$event-live:example.com".try_into().unwrap();
-        let mut previous = HashMap::new();
-        let mut previous_state = make_state("hello");
-        previous_state.advance_displayed(3);
-        previous.insert(event_id.clone(), previous_state);
-
-        let (rebuilt, should_schedule_frame) = rebuild_streaming_messages_for_full_snapshot(
-            [(event_id.clone(), String::from("hello world"), true)],
-            Some(&previous),
-        );
-
-        let restored = rebuilt.get(&event_id).unwrap();
-        assert_eq!(restored.displayed_char_count, 3);
-        assert!(restored.is_live);
-        assert!(should_schedule_frame);
-    }
-
-    #[test]
-    fn test_full_snapshot_rebuild_skips_live_without_cached_state() {
-        // Without previous state, full-snapshot rebuild must NOT create new
-        // animations — the SDK may not have aggregated edits yet, so
-        // completed messages can still appear as `live`.
-        let event_id: OwnedEventId = "$event-live:example.com".try_into().unwrap();
-
-        let (rebuilt, should_schedule_frame) = rebuild_streaming_messages_for_full_snapshot(
-            [(event_id.clone(), String::from("hello world"), true)],
-            None,
-        );
-
-        assert!(rebuilt.is_empty());
-        assert!(!should_schedule_frame);
     }
 
     #[test]
@@ -16213,134 +13603,6 @@ mod tests {
     #[test]
     fn bot_badge_text_is_centered_within_badge() {
         assert!(bot_badge_label_center_y() < (BOT_BADGE_HEIGHT * 0.5));
-    }
-
-    #[test]
-    fn test_bot_detection_configured_parent() {
-        let user_id: OwnedUserId = "@octosbot:127.0.0.1:8128".try_into().unwrap();
-        let resolved_parent_bot_user_id = Some(user_id.clone());
-        let known_bot_user_ids = Vec::new();
-
-        assert!(is_known_or_likely_bot(
-            user_id.as_ref(),
-            resolved_parent_bot_user_id.as_deref(),
-            &known_bot_user_ids,
-        ));
-    }
-
-    #[test]
-    fn test_bot_detection_heuristic_fallback() {
-        let user_id: OwnedUserId = "@myservice_bot:other.server".try_into().unwrap();
-        let known_bot_user_ids = Vec::new();
-
-        assert!(is_known_or_likely_bot(
-            user_id.as_ref(),
-            None,
-            &known_bot_user_ids,
-        ));
-    }
-
-    #[test]
-    fn test_bot_detection_child_bot() {
-        let user_id: OwnedUserId = "@octosbot_weather:127.0.0.1:8128".try_into().unwrap();
-        let known_bot_user_ids = vec![user_id.clone()];
-
-        assert!(is_known_or_likely_bot(
-            user_id.as_ref(),
-            None,
-            &known_bot_user_ids,
-        ));
-    }
-
-    #[test]
-    fn test_bot_detection_rejects_normal_user() {
-        let user_id: OwnedUserId = "@alice:127.0.0.1:8128".try_into().unwrap();
-        let known_bot_user_ids = Vec::new();
-
-        assert!(!is_known_or_likely_bot(
-            user_id.as_ref(),
-            None,
-            &known_bot_user_ids,
-        ));
-    }
-
-    #[test]
-    fn test_timeline_bot_detection_uses_room_bot_user_ids() {
-        let user_id: OwnedUserId = "@octosbot_bob:127.0.0.1:8128".try_into().unwrap();
-        let room_bot_user_ids = vec![user_id.clone()];
-        let known_bot_user_ids = Vec::new();
-
-        assert!(is_timeline_sender_bot(
-            user_id.as_ref(),
-            None,
-            &room_bot_user_ids,
-            &known_bot_user_ids,
-        ));
-    }
-
-    #[test]
-    fn test_timeline_bot_context_cache_matches_its_full_identity_fingerprint() {
-        let room_id: OwnedRoomId = "!room:example.org".try_into().unwrap();
-        let known_bot_user_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let cached = CachedTimelineBotContext {
-            room_id: room_id.clone(),
-            room_members: None,
-            app_service_enabled: false,
-            room_is_bound: false,
-            persisted_bound_bot_user_id: None,
-            persisted_bound_bot_user_ids: Vec::new(),
-            resolved_parent_bot_user_id: None,
-            known_bot_user_ids: vec![known_bot_user_id.clone()],
-            value: TimelineBotContext::default(),
-        };
-
-        assert!(cached.matches(
-            &room_id,
-            None,
-            false,
-            false,
-            None,
-            &[],
-            None,
-            &[known_bot_user_id],
-        ));
-        assert!(!cached.matches(
-            &room_id,
-            None,
-            false,
-            false,
-            None,
-            &[],
-            None,
-            &[],
-        ));
-    }
-
-    #[test]
-    fn test_registry_agent_detected_as_bot_sender() {
-        // An agent known only via the global AgentRegistry (not the app-service
-        // known-bot list, and with a non-bot-like localpart) is still detected.
-        let agent_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-        app_state
-            .agent_registry
-            .register(agent_id.clone(), crate::app::AgentEntry::default());
-
-        let known_bot_user_ids = timeline_known_bot_user_ids(&app_state);
-        assert!(is_known_or_likely_bot(agent_id.as_ref(), None, &known_bot_user_ids));
-    }
-
-    #[test]
-    fn test_room_props_known_bot_user_ids_include_registry_agents() {
-        let agent_id: OwnedUserId = "@agent:example.org".try_into().unwrap();
-        let mut app_state = AppState::default();
-        app_state
-            .agent_registry
-            .register(agent_id.clone(), crate::app::AgentEntry::default());
-
-        let known_bot_user_ids = timeline_known_bot_user_ids(&app_state);
-
-        assert!(known_bot_user_ids.iter().any(|id| id == &agent_id));
     }
 
     #[test]
@@ -16456,26 +13718,6 @@ mod tests {
     }
 
     #[test]
-    fn test_detected_bot_binding_uses_registry_augmented_known_bots() {
-        let src = include_str!("mod.rs");
-        let fn_pos = src
-            .find("fn detected_bot_binding_for_members")
-            .expect("detected_bot_binding_for_members should exist");
-        let fn_src = &src[fn_pos..src[fn_pos..].find("fn is_likely_bot_user_id")
-            .map(|end| fn_pos + end)
-            .unwrap_or(src.len())];
-
-        assert!(
-            fn_src.contains("timeline_known_bot_user_ids(app_state)"),
-            "DM bot binding detection should include AgentRegistry agents such as OctosDirect",
-        );
-        assert!(
-            !fn_src.contains("app_state.bot_settings.known_bot_user_ids()"),
-            "DM bot binding detection must not read only raw AppService known-bots",
-        );
-    }
-
-    #[test]
     fn test_room_info_title_bot_pill_shown_when_member_is_registered_agent() {
         let room_id: OwnedRoomId = "!group:example.org".try_into().unwrap();
         let agent_id: OwnedUserId = "@octos_mac:example.org".try_into().unwrap();
@@ -16586,1205 +13828,4 @@ mod tests {
         );
     }
 
-    #[test]
-    fn test_non_agent_user_not_detected_as_bot() {
-        // Empty registry, empty known-bot list, app-service disabled.
-        let app_state = AppState::default();
-        let human_id: OwnedUserId = "@human:example.org".try_into().unwrap();
-
-        let known_bot_user_ids = timeline_known_bot_user_ids(&app_state);
-        assert!(!is_known_or_likely_bot(human_id.as_ref(), None, &known_bot_user_ids));
-    }
-
-    #[test]
-    fn test_empty_registry_and_no_known_bots_shows_no_bot_card() {
-        let app_state = AppState::default();
-        let known_bot_user_ids = timeline_known_bot_user_ids(&app_state);
-        assert!(known_bot_user_ids.is_empty());
-
-        let sender_id: OwnedUserId = "@someone:example.org".try_into().unwrap();
-        let is_bot_sender = is_known_or_likely_bot(sender_id.as_ref(), None, &known_bot_user_ids);
-        let render_state = compute_bot_timeline_render_state("hello", is_bot_sender);
-        assert!(!render_state.show_card);
-    }
-
-    #[test]
-    fn test_listbots_parser_ignores_octos_service_urls_and_ports() {
-        let parsed = extract_bot_user_ids_from_listbots_reply(
-            "Octos service: http://127.0.0.1:8787\nKnown bots: @octosbot:example.org",
-            None,
-        );
-
-        assert_eq!(
-            parsed,
-            vec!["@octosbot:example.org".parse::<OwnedUserId>().unwrap()],
-        );
-        assert!(
-            parsed.iter().all(|user_id| user_id.localpart() != ""),
-            "service URL port fragments must not become Matrix user IDs",
-        );
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_extracts_status_provider_body_and_footer() {
-        let body = "施法中\nvia moonshot@api (kimi-k2.5)\n\n你好！我是 **Alex**\n\n_moonshot@api/kimi-k2.5 · 5.3K in · 330 out · 6s_";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers.status.as_deref(), Some("施法中"));
-        assert_eq!(layers.provider.as_deref(), Some("via moonshot@api (kimi-k2.5)"));
-        assert_eq!(layers.body, "你好！我是 **Alex**");
-        assert_eq!(
-            layers.footer.as_deref(),
-            Some("_moonshot@api/kimi-k2.5 · 5.3K in · 330 out · 6s_"),
-        );
-    }
-
-    #[test]
-    fn test_parse_octos_actions_skips_malformed_entries() {
-        let actions = parse_octos_actions_from_content(&serde_json::json!({
-            "org.octos.actions": [
-                { "id": "retry_pptx", "label": "Regenerate PPT", "style": "primary" },
-                { "label": "Missing id" },
-                { "id": "cancel", "label": "Cancel", "style": "secondary" }
-            ]
-        }));
-
-        assert_eq!(actions.len(), 2);
-        assert_eq!(actions[0].id, "retry_pptx");
-        assert_eq!(actions[1].id, "cancel");
-    }
-
-    #[test]
-    fn test_parse_octos_actions_truncates_after_six() {
-        let actions = parse_octos_actions_from_content(&serde_json::json!({
-            "org.octos.actions": [
-                { "id": "a1", "label": "A1" },
-                { "id": "a2", "label": "A2" },
-                { "id": "a3", "label": "A3" },
-                { "id": "a4", "label": "A4" },
-                { "id": "a5", "label": "A5" },
-                { "id": "a6", "label": "A6" },
-                { "id": "a7", "label": "A7" }
-            ]
-        }));
-
-        assert_eq!(actions.len(), 6);
-        assert_eq!(actions.last().map(|action| action.id.as_str()), Some("a6"));
-    }
-
-    #[test]
-    fn test_parse_octos_actions_reads_m_new_content_wrapper() {
-        let actions = parse_octos_actions_from_content(&serde_json::json!({
-            "m.new_content": {
-                "org.octos.actions": [
-                    { "id": "confirm", "label": "确认", "style": "primary" },
-                    { "id": "cancel", "label": "取消", "style": "secondary" }
-                ]
-            },
-            "org.octos.actions": [
-                { "id": "stale", "label": "旧按钮" }
-            ]
-        }));
-
-        assert_eq!(actions.len(), 2);
-        assert_eq!(actions[0].id, "confirm");
-        assert_eq!(actions[1].id, "cancel");
-    }
-
-    const TEST_AGENTCHAT_AGENT: &str = "test_agent";
-    const TEST_AGENTCHAT_PROJECT: &str = "test_project";
-    const TEST_AGENTCHAT_PROJECT_ROOM_ID: &str = "!project:example.test";
-
-    fn valid_agentchat_approval_content() -> serde_json::Value {
-        serde_json::json!({
-            "msgtype": "com.agentchat.approval.request.v1",
-            "body": format!("Approval required for {TEST_AGENTCHAT_AGENT}"),
-            "com.agentchat.approval": {
-                "version": 1,
-                "kind": "request",
-                "agent": TEST_AGENTCHAT_AGENT,
-                "project": TEST_AGENTCHAT_PROJECT,
-                "project_room_id": TEST_AGENTCHAT_PROJECT_ROOM_ID,
-                "request_id": "approval_0123456789abcdef0123456789abcdef",
-                "upstream_request_id": "turn-1:Bash",
-                "input_digest": "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
-                "runtime": "codex",
-                "tool_name": "Bash",
-                "description": "Create a GitHub issue",
-                "input_preview": "gh issue create --title test",
-                "expires_at": 1784745600000u64,
-                "actions": [
-                    { "id": "approve_once", "label": "Approve once", "style": "primary" },
-                    { "id": "deny", "label": "Deny", "style": "danger" }
-                ]
-            }
-        })
-    }
-
-    #[test]
-    fn test_parse_agentchat_owner_approval_request() {
-        let content = valid_agentchat_approval_content();
-        let payload = parse_octos_action_payload_for_render(Some(&content), Some(&content));
-        let approval = payload.approval_request.expect("agent-chat approval should parse");
-
-        assert_eq!(approval.request_id, "approval_0123456789abcdef0123456789abcdef");
-        assert_eq!(approval.tool_args_digest, "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef");
-        assert_eq!(payload.actions.iter().map(|action| action.id.as_str()).collect::<Vec<_>>(), vec!["approve_once", "deny"]);
-        assert!(!payload.malformed_approval_request);
-        assert!(matches!(
-            approval.protocol,
-            ApprovalProtocol::AgentChat { ref agent, ref project, ref project_room_id }
-                if agent == TEST_AGENTCHAT_AGENT
-                    && project == TEST_AGENTCHAT_PROJECT
-                    && project_room_id == TEST_AGENTCHAT_PROJECT_ROOM_ID
-        ));
-    }
-
-    #[test]
-    fn test_agentchat_approval_buttons_expire_at_deadline() {
-        let content = valid_agentchat_approval_content();
-        let payload = parse_octos_action_payload_for_render(Some(&content), Some(&content));
-        let approval = payload.approval_request.as_ref().expect("agent-chat approval should parse");
-        let expires_at = approval.expires_at.parse::<u64>().unwrap();
-        let current_user_id = UserId::parse("@owner:example.test").unwrap();
-
-        let live = compute_action_button_render_state_at(
-            &payload.actions,
-            Some(approval),
-            Some(current_user_id.as_ref()),
-            expires_at - 1,
-        );
-        assert!(live.buttons_enabled);
-        assert_eq!(live.approval_card.as_ref().map(|card| card.expired), Some(false));
-
-        let expired = compute_action_button_render_state_at(
-            &payload.actions,
-            Some(approval),
-            Some(current_user_id.as_ref()),
-            expires_at,
-        );
-        assert!(!expired.buttons_enabled);
-        assert_eq!(expired.approval_card.as_ref().map(|card| card.expired), Some(true));
-    }
-
-    fn agentchat_approval_button_request(expires_at: &str) -> OctosActionButtonRequest {
-        OctosActionButtonRequest::Approval {
-            protocol: ApprovalProtocol::AgentChat {
-                agent: TEST_AGENTCHAT_AGENT.into(),
-                project: TEST_AGENTCHAT_PROJECT.into(),
-                project_room_id: TEST_AGENTCHAT_PROJECT_ROOM_ID.into(),
-            },
-            request_id: "approval_0123456789abcdef0123456789abcdef".into(),
-            title: "Run command".into(),
-            decision: "approve_once".into(),
-            label: "Approve once".into(),
-            tool_args_digest: "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef".into(),
-            expires_at: expires_at.into(),
-            style: OctosActionStyle::Primary,
-        }
-    }
-
-    #[test]
-    fn test_approval_expiry_timer_uses_earliest_visible_deadline() {
-        let generic = OctosActionButtonRequest::Generic {
-            action_id: "retry".into(),
-            label: "Retry".into(),
-            style: OctosActionStyle::Secondary,
-        };
-        let later = agentchat_approval_button_request("1500");
-        let sooner = agentchat_approval_button_request("1250");
-
-        assert_eq!(
-            next_approval_expiry_timeout([&generic, &later, &sooner], 1000),
-            Some(Duration::from_millis(250)),
-        );
-        assert_eq!(
-            next_approval_expiry_timeout([&sooner], 1250),
-            Some(Duration::from_millis(1)),
-        );
-        assert_eq!(
-            next_approval_expiry_timeout([&generic], 1000),
-            None,
-        );
-    }
-
-    #[test]
-    fn test_agentchat_public_status_has_no_actions() {
-        let body = format!("Agent {TEST_AGENTCHAT_AGENT} is waiting for approval from its owner.");
-        let content = serde_json::json!({
-            "msgtype": "com.agentchat.approval.status.v1",
-            "body": body.clone(),
-            "com.agentchat.approval": {
-                "version": 1,
-                "kind": "status",
-                "agent": TEST_AGENTCHAT_AGENT,
-                "project": TEST_AGENTCHAT_PROJECT,
-                "state": "waiting_for_owner"
-            }
-        });
-        let payload = parse_octos_action_payload_for_render(Some(&content), Some(&content));
-
-        assert_eq!(
-            agentchat_custom_message_body_from_content(&content),
-            Some(body.as_str()),
-        );
-        assert!(payload.approval_request.is_none());
-        assert!(payload.actions.is_empty());
-        assert!(!payload.malformed_approval_request);
-    }
-
-    #[test]
-    fn test_malformed_agentchat_owner_approval_request_hides_buttons() {
-        let mut content = valid_agentchat_approval_content();
-        content["com.agentchat.approval"]["input_digest"] = serde_json::json!("not-a-digest");
-        let payload = parse_octos_action_payload_for_render(Some(&content), Some(&content));
-        let state = compute_action_button_render_state(&payload.actions, payload.approval_request.as_ref(), None);
-
-        assert!(payload.malformed_approval_request);
-        assert!(payload.actions.is_empty());
-        assert!(!state.show_container);
-    }
-
-    #[test]
-    fn test_build_agentchat_approval_verdict() {
-        let timeline_kind = TimelineKind::MainRoom {
-            room_id: "!approval:example.test".try_into().unwrap(),
-        };
-        let source_event_id: OwnedEventId = "$approval-request".try_into().unwrap();
-        let original_sender: OwnedUserId = "@agent-bridge:example.test".try_into().unwrap();
-        let request = build_agentchat_approval_verdict_request(
-            &timeline_kind,
-            "Approve once",
-            "approval_0123456789abcdef0123456789abcdef",
-            "approve_once",
-            "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef",
-            TEST_AGENTCHAT_AGENT,
-            TEST_AGENTCHAT_PROJECT,
-            TEST_AGENTCHAT_PROJECT_ROOM_ID,
-            source_event_id.as_ref(),
-            original_sender.as_ref(),
-        );
-
-        let verdict = &request.content["com.agentchat.approval"];
-        assert_eq!(request.content["msgtype"], AGENTCHAT_APPROVAL_VERDICT_MSGTYPE);
-        assert_eq!(verdict["kind"], "verdict");
-        assert_eq!(verdict["action"], "approve_once");
-        assert_eq!(verdict["agent"], TEST_AGENTCHAT_AGENT);
-        assert_eq!(verdict["project"], TEST_AGENTCHAT_PROJECT);
-        assert_eq!(verdict["project_room_id"], TEST_AGENTCHAT_PROJECT_ROOM_ID);
-        assert_eq!(verdict["request_id"], "approval_0123456789abcdef0123456789abcdef");
-        assert_eq!(verdict["input_digest"], "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef");
-        assert_eq!(request.target_user_id, original_sender);
-    }
-
-    #[test]
-    fn test_agentchat_approval_uses_original_content() {
-        let original = valid_agentchat_approval_content();
-        let mut edited = original.clone();
-        edited["com.agentchat.approval"]["request_id"] = serde_json::json!("approval_ffffffffffffffffffffffffffffffff");
-        edited["com.agentchat.approval"]["input_digest"] = serde_json::json!("ffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffffff");
-
-        let payload = parse_octos_action_payload_for_render(Some(&edited), Some(&original));
-        let approval = payload.approval_request.expect("original approval should parse");
-        assert_eq!(approval.request_id, "approval_0123456789abcdef0123456789abcdef");
-        assert_eq!(approval.tool_args_digest, "0123456789abcdef0123456789abcdef0123456789abcdef0123456789abcdef");
-    }
-
-    #[test]
-    fn test_parse_octos_approval_request_from_content() {
-        let approval = parse_octos_approval_request_from_content(&serde_json::json!({
-            "org.octos.approval_request": {
-                "request_id": "req_abc123",
-                "tool_name": "shell",
-                "tool_args_digest": "sha256:4bf5",
-                "title": "Execute shell command",
-                "summary": "rm -rf ~/tmp/cache",
-                "risk_level": "critical",
-                "authorized_approvers": ["@alice:example.org"],
-                "expires_at": "2026-04-14T14:30:00Z",
-                "on_timeout": "notify"
-            }
-        })).expect("approval request should parse");
-
-        assert_eq!(approval.request_id, "req_abc123");
-        assert_eq!(approval.tool_name, "shell");
-        assert_eq!(approval.tool_args_digest, "sha256:4bf5");
-        assert_eq!(approval.title, "Execute shell command");
-        assert_eq!(approval.summary, "rm -rf ~/tmp/cache");
-        assert_eq!(approval.risk_level, OctosApprovalRiskLevel::Critical);
-        assert_eq!(approval.authorized_approvers, vec!["@alice:example.org"]);
-        assert_eq!(approval.on_timeout, OctosApprovalTimeoutBehavior::Notify);
-    }
-
-    #[test]
-    fn test_parse_octos_approval_request_ignores_m_new_content_wrapper() {
-        let approval = parse_octos_approval_request_from_content(&serde_json::json!({
-            "org.octos.approval_request": {
-                "request_id": "req_original",
-                "tool_name": "shell",
-                "tool_args_digest": "sha256:4bf5",
-                "title": "Original request",
-                "summary": "rm -rf ~/tmp/cache",
-                "risk_level": "critical",
-                "authorized_approvers": ["@alice:example.org"],
-                "expires_at": "2026-04-14T14:30:00Z",
-                "on_timeout": "notify"
-            },
-            "m.new_content": {
-                "org.octos.approval_request": {
-                    "request_id": "req_edited",
-                    "tool_name": "shell",
-                    "tool_args_digest": "sha256:mallory",
-                    "title": "Edited request",
-                    "summary": "whoami",
-                    "risk_level": "normal",
-                    "authorized_approvers": ["@mallory:example.org"],
-                    "expires_at": "2026-04-14T14:30:00Z",
-                    "on_timeout": "notify"
-                }
-            }
-        })).expect("approval request should parse from original content");
-
-        assert_eq!(approval.request_id, "req_original");
-        assert_eq!(approval.authorized_approvers, vec!["@alice:example.org"]);
-        assert_eq!(approval.risk_level, OctosApprovalRiskLevel::Critical);
-    }
-
-    #[test]
-    fn test_parse_octos_approval_request_rejects_empty_authorized_approvers() {
-        assert!(parse_octos_approval_request_from_content(&serde_json::json!({
-            "org.octos.approval_request": {
-                "request_id": "req_abc123",
-                "tool_name": "shell",
-                "tool_args_digest": "sha256:4bf5",
-                "title": "Execute shell command",
-                "summary": "rm -rf ~/tmp/cache",
-                "risk_level": "critical",
-                "authorized_approvers": [],
-                "expires_at": "2026-04-14T14:30:00Z",
-                "on_timeout": "notify"
-            }
-        })).is_none());
-    }
-
-    #[test]
-    fn test_build_approval_response_request_targets_original_sender() {
-        let timeline_kind = TimelineKind::MainRoom {
-            room_id: "!room:127.0.0.1:8128".try_into().unwrap(),
-        };
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let original_sender: OwnedUserId = "@octosbot:127.0.0.1:8128".try_into().unwrap();
-
-        let request = build_octos_approval_response_request(
-            &timeline_kind,
-            "Execute shell command",
-            "req_abc123",
-            "approve",
-            "sha256:4bf5",
-            source_event_id.as_ref(),
-            original_sender.as_ref(),
-        );
-
-        assert_eq!(request.timeline_kind, timeline_kind);
-        assert_eq!(request.target_user_id, original_sender);
-        assert!(!request.explicit_room);
-        assert_eq!(request.content["org.octos.approval_response"]["request_id"], "req_abc123");
-        assert_eq!(request.content["org.octos.approval_response"]["decision"], "approve");
-        assert_eq!(request.content["org.octos.approval_response"]["tool_args_digest"], "sha256:4bf5");
-    }
-
-    #[test]
-    fn test_action_buttons_render_state_hidden_without_actions() {
-        let state = compute_action_button_render_state(&[], None, None);
-
-        assert!(!state.show_container);
-        assert!(state.visible_slots.is_empty());
-    }
-
-    #[test]
-    fn test_action_buttons_render_state_with_primary_secondary_danger() {
-        let state = compute_action_button_render_state(&[
-            OctosActionButton {
-                id: "retry".into(),
-                label: "Regenerate PPT".into(),
-                style: OctosActionStyle::Primary,
-            },
-            OctosActionButton {
-                id: "cancel".into(),
-                label: "Cancel".into(),
-                style: OctosActionStyle::Secondary,
-            },
-            OctosActionButton {
-                id: "delete".into(),
-                label: "Delete".into(),
-                style: OctosActionStyle::Danger,
-            },
-        ], None, None);
-
-        assert!(state.show_container);
-        assert!(state.show_button_row);
-        assert!(state.buttons_enabled);
-        assert!(state.approval_card.is_none());
-        assert_eq!(state.visible_slots.len(), 3);
-        assert_eq!(state.visible_slots[0].style, OctosActionStyle::Primary);
-        assert_eq!(state.visible_slots[1].style, OctosActionStyle::Secondary);
-        assert_eq!(state.visible_slots[2].style, OctosActionStyle::Danger);
-    }
-
-    #[test]
-    fn test_approval_buttons_disabled_for_unauthorized_user() {
-        let approval_request = OctosApprovalRequest {
-            protocol: ApprovalProtocol::Octos,
-            request_id: "req_abc123".into(),
-            tool_name: "shell".into(),
-            tool_args_digest: "sha256:4bf5".into(),
-            title: "Execute shell command".into(),
-            summary: "rm -rf ~/tmp/cache".into(),
-            risk_level: OctosApprovalRiskLevel::Critical,
-            authorized_approvers: vec!["@alice:example.org".into()],
-            expires_at: "2026-04-14T14:30:00Z".into(),
-            on_timeout: OctosApprovalTimeoutBehavior::Notify,
-        };
-        let current_user_id = UserId::parse("@mallory:example.org").unwrap();
-        let state = compute_action_button_render_state(&[
-            OctosActionButton {
-                id: "approve".into(),
-                label: "Approve".into(),
-                style: OctosActionStyle::Primary,
-            },
-            OctosActionButton {
-                id: "deny".into(),
-                label: "Deny".into(),
-                style: OctosActionStyle::Danger,
-            },
-        ], Some(&approval_request), Some(current_user_id.as_ref()));
-
-        assert!(state.show_container);
-        assert!(state.show_button_row);
-        assert!(!state.buttons_enabled);
-        assert_eq!(
-            state.approval_card.as_ref().map(|card| card.title.as_str()),
-            Some("Execute shell command"),
-        );
-        assert_eq!(
-            state.approval_card.as_ref().map(|card| card.summary.as_str()),
-            Some("rm -rf ~/tmp/cache"),
-        );
-    }
-
-    #[test]
-    fn test_selected_action_reduces_visible_slots_to_clicked_button() {
-        let render_state = compute_action_button_render_state(&[
-            OctosActionButton {
-                id: "approve".into(),
-                label: "Approve".into(),
-                style: OctosActionStyle::Primary,
-            },
-            OctosActionButton {
-                id: "deny".into(),
-                label: "Deny".into(),
-                style: OctosActionStyle::Danger,
-            },
-        ], None, None);
-
-        let visible_slots = action_button_render_slots_for_display(&render_state, Some(&SelectedOctosActionState {
-            id: "deny".into(),
-            label: "Deny".into(),
-            style: OctosActionStyle::Danger,
-        }));
-
-        assert_eq!(visible_slots.len(), 1);
-        assert_eq!(visible_slots[0].id, "deny");
-        assert_eq!(visible_slots[0].label, "✓ Deny");
-        assert_eq!(visible_slots[0].style, OctosActionStyle::Danger);
-    }
-
-    #[test]
-    fn test_generic_actions_without_approval_request_remain_supported() {
-        let payload = parse_octos_action_payload_for_render(
-            Some(&serde_json::json!({
-                "org.octos.actions": [
-                    { "id": "retry_pptx", "label": "Regenerate PPT", "style": "primary" }
-                ]
-            })),
-            None,
-        );
-
-        assert!(payload.approval_request.is_none());
-        assert!(!payload.malformed_approval_request);
-        assert_eq!(payload.actions.len(), 1);
-        assert_eq!(payload.actions[0].id, "retry_pptx");
-    }
-
-    #[test]
-    fn test_malformed_approval_request_hides_buttons() {
-        let payload = parse_octos_action_payload_for_render(
-            Some(&serde_json::json!({
-                "org.octos.actions": [
-                    { "id": "approve", "label": "Approve", "style": "primary" },
-                    { "id": "deny", "label": "Deny", "style": "danger" }
-                ]
-            })),
-            Some(&serde_json::json!({
-                "org.octos.approval_request": {
-                    "request_id": "req_abc123"
-                },
-                "org.octos.actions": [
-                    { "id": "approve", "label": "Approve", "style": "primary" },
-                    { "id": "deny", "label": "Deny", "style": "danger" }
-                ]
-            })),
-        );
-        let state = compute_action_button_render_state(
-            &payload.actions,
-            payload.approval_request.as_ref(),
-            None,
-        );
-
-        assert!(payload.malformed_approval_request);
-        assert!(!state.show_container);
-        assert!(state.visible_slots.is_empty());
-    }
-
-    #[test]
-    fn test_approval_request_ignores_m_replace_edits() {
-        let payload = parse_octos_action_payload_for_render(
-            Some(&serde_json::json!({
-                "m.new_content": {
-                    "org.octos.approval_request": {
-                        "request_id": "req_replaced",
-                        "tool_name": "shell",
-                        "tool_args_digest": "sha256:replaced",
-                        "title": "Replaced request",
-                        "summary": "echo hacked",
-                        "risk_level": "normal",
-                        "authorized_approvers": ["@mallory:example.org"],
-                        "expires_at": "2026-04-14T14:35:00Z",
-                        "on_timeout": "notify"
-                    },
-                    "org.octos.actions": [
-                        { "id": "approve", "label": "Approve", "style": "primary" },
-                        { "id": "deny", "label": "Deny", "style": "danger" }
-                    ]
-                }
-            })),
-            Some(&serde_json::json!({
-                "org.octos.approval_request": {
-                    "request_id": "req_original",
-                    "tool_name": "shell",
-                    "tool_args_digest": "sha256:original",
-                    "title": "Original request",
-                    "summary": "rm -rf ~/tmp/cache",
-                    "risk_level": "critical",
-                    "authorized_approvers": ["@alice:example.org"],
-                    "expires_at": "2026-04-14T14:30:00Z",
-                    "on_timeout": "notify"
-                },
-                "org.octos.actions": [
-                    { "id": "approve", "label": "Approve", "style": "primary" },
-                    { "id": "deny", "label": "Deny", "style": "danger" }
-                ]
-            })),
-        );
-        let current_user_id = UserId::parse("@alice:example.org").unwrap();
-        let state = compute_action_button_render_state_at(
-            &payload.actions,
-            payload.approval_request.as_ref(),
-            Some(current_user_id.as_ref()),
-            0,
-        );
-
-        assert_eq!(
-            payload.approval_request.as_ref().map(|approval| approval.request_id.as_str()),
-            Some("req_original"),
-        );
-        assert!(state.buttons_enabled);
-        assert_eq!(
-            state.approval_card.as_ref().map(|card| card.title.as_str()),
-            Some("Original request"),
-        );
-    }
-
-    #[test]
-    fn test_build_action_response_request_targets_original_sender() {
-        let timeline_kind = TimelineKind::MainRoom {
-            room_id: "!room:127.0.0.1:8128".try_into().unwrap(),
-        };
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let original_sender: OwnedUserId = "@octosbot_weather:127.0.0.1:8128".try_into().unwrap();
-
-        let request = build_octos_action_response_request(
-            &timeline_kind,
-            "Regenerate PPT",
-            "retry_pptx",
-            source_event_id.as_ref(),
-            original_sender.as_ref(),
-        );
-
-        assert_eq!(request.timeline_kind, timeline_kind);
-        assert_eq!(request.target_user_id, original_sender);
-        assert!(!request.explicit_room);
-    }
-
-    #[test]
-    fn test_build_action_response_request_preserves_reply_relation() {
-        let timeline_kind = TimelineKind::MainRoom {
-            room_id: "!room:127.0.0.1:8128".try_into().unwrap(),
-        };
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let original_sender: OwnedUserId = "@octosbot_weather:127.0.0.1:8128".try_into().unwrap();
-
-        let request = build_octos_action_response_request(
-            &timeline_kind,
-            "Regenerate PPT",
-            "retry_pptx",
-            source_event_id.as_ref(),
-            original_sender.as_ref(),
-        );
-
-        let action_response = &request.content["org.octos.action_response"];
-        assert_eq!(request.content["body"], "[Action: Regenerate PPT]");
-        assert_eq!(action_response["action_id"], "retry_pptx");
-        assert_eq!(action_response["source_event_id"], "$orig123");
-        assert_eq!(request.content["m.relates_to"]["m.in_reply_to"]["event_id"], "$orig123");
-    }
-
-    #[test]
-    fn test_disable_action_buttons_marks_source_event_disabled() {
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let mut disabled = HashSet::new();
-
-        mark_action_buttons_disabled(&mut disabled, &source_event_id);
-
-        assert!(are_action_buttons_disabled(&disabled, source_event_id.as_ref()));
-    }
-
-    #[test]
-    fn test_reenable_action_buttons_clears_disabled_state() {
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let mut disabled = HashSet::new();
-        mark_action_buttons_disabled(&mut disabled, &source_event_id);
-
-        clear_action_buttons_disabled(&mut disabled, source_event_id.as_ref());
-
-        assert!(!are_action_buttons_disabled(&disabled, source_event_id.as_ref()));
-    }
-
-    #[test]
-    fn test_selected_action_state_marks_and_clears_by_source_event_id() {
-        let source_event_id: OwnedEventId = "$orig123".try_into().unwrap();
-        let mut selected_actions = HashMap::new();
-
-        mark_selected_octos_action(
-            &mut selected_actions,
-            &source_event_id,
-            "approve",
-            "Approve",
-            OctosActionStyle::Primary,
-        );
-        assert_eq!(
-            selected_actions.get(&source_event_id).map(|state| state.label.as_str()),
-            Some("Approve"),
-        );
-
-        clear_selected_octos_action(&mut selected_actions, source_event_id.as_ref());
-        assert!(!selected_actions.contains_key(&source_event_id));
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_extracts_footer_without_provider_prefix() {
-        let body = "PPT 已经生成并发送了！\n\n你应该已经收到了文件。\n\n_moonshot@api/kimi-k2.5 · 11.0K in · 279 out · 9s_";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers.status, None);
-        assert_eq!(layers.provider, None);
-        assert_eq!(layers.body, "PPT 已经生成并发送了！\n\n你应该已经收到了文件。");
-        assert_eq!(
-            layers.footer.as_deref(),
-            Some("_moonshot@api/kimi-k2.5 · 11.0K in · 279 out · 9s_"),
-        );
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_falls_back_for_unmatched_bot_text() {
-        let body = "你好！我是 Alex。\n今天可以帮你查天气。";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers, BotTimelineLayers::plain(body));
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_ignores_regular_user_messages() {
-        let body = "via moonshot@api (kimi-k2.5)\n\n这不是 bot 消息。";
-
-        let layers = parse_bot_timeline_layers(body, false);
-
-        assert_eq!(layers, BotTimelineLayers::plain(body));
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_prefers_safe_fallback_for_malformed_metadata() {
-        let body = "施法中\n这个不是 provider 行\n\n你好，我还在。";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers, BotTimelineLayers::plain(body));
-    }
-
-    #[test]
-    #[ignore = "pre-existing failure on main (1.0.0-alpha.1): parses status/provider layers instead of treating the metadata as invalid. See issues/011."]
-    fn test_parse_bot_timeline_layers_invalid_metadata_does_not_panic() {
-        let body = "施法中\nvia moonshot@api (kimi-k2.5)\n\n_\n";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers, BotTimelineLayers::plain(body));
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_tolerates_streaming_cursor_in_footer() {
-        let body = "via moonshot@api (kimi-k2.5)\n\n你好！我是 **Alex**\n\n_moonshot@api/kimi-k2.5 · 5.3K in · 330 out · 6s_ ●";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers.body, "你好！我是 **Alex**");
-        assert_eq!(
-            layers.footer.as_deref(),
-            Some("_moonshot@api/kimi-k2.5 · 5.3K in · 330 out · 6s_"),
-        );
-    }
-
-    #[test]
-    fn test_parse_bot_timeline_layers_promotes_metrics_only_body_to_footer() {
-        let body = "疯狂输出中\nvia moonshot@api (kimi-k2.5)\n4s";
-
-        let layers = parse_bot_timeline_layers(body, true);
-
-        assert_eq!(layers.status.as_deref(), Some("疯狂输出中"));
-        assert_eq!(layers.provider.as_deref(), Some("via moonshot@api (kimi-k2.5)"));
-        assert!(layers.body.is_empty());
-        assert_eq!(layers.footer.as_deref(), Some("4s"));
-    }
-
-    #[test]
-    fn test_rich_markdown_streaming_prefers_full_snapshot_rendering() {
-        let formatted = FormattedBody::html("<p><strong>OpenClaw</strong></p>");
-        assert!(should_render_streaming_full_snapshot(
-            "根据搜索结果， **OpenClaw** 有两个不同的项目。",
-            Some(&formatted),
-            true,
-        ));
-    }
-
-    #[test]
-    fn test_plain_text_streaming_keeps_typewriter_path() {
-        assert!(!should_render_streaming_full_snapshot(
-            "你好，我是 Octos。",
-            None,
-            true,
-        ));
-    }
-
-    #[test]
-    fn test_bot_timeline_card_visible_for_bot_text_message() {
-        let state = compute_bot_timeline_render_state(
-            "施法中\nvia moonshot@api (kimi-k2.5)\n\n你好！我是 Alex。\n\n_moonshot@api/kimi-k2.5 · 1.2K in · 88 out · 2s_",
-            true,
-        );
-
-        assert!(state.show_card);
-        assert_eq!(state.body, "你好！我是 Alex。");
-    }
-
-    #[test]
-    fn test_bot_timeline_card_hidden_for_regular_user_message() {
-        let state = compute_bot_timeline_render_state("你好", false);
-
-        assert!(!state.show_card);
-    }
-
-    #[test]
-    fn test_bot_status_strip_renders_above_body_and_not_inside_body() {
-        let state = compute_bot_timeline_render_state(
-            "施法中\nvia moonshot@api (kimi-k2.5)\n\n你好！我是 Alex。",
-            true,
-        );
-
-        assert_eq!(state.status.as_deref(), Some("施法中"));
-        assert!(state.show_status_strip);
-        assert!(!state.body.starts_with("施法中"));
-    }
-
-    #[test]
-    fn test_agent_chat_puppets_are_treated_as_bots() {
-        // Real MXIDs from an agent-chat team: none contain "bot".
-        for localpart in [
-            "ac_tyrese_coordinator",
-            "ac_tyrese_implementer",
-            "ac_tyrese_reviewer",
-            "ac_tyrese_final_reviewer",
-            "wf_coordinator",
-        ] {
-            assert!(
-                is_agent_chat_puppet_localpart(localpart),
-                "{localpart} must be recognised as an agent-chat puppet"
-            );
-        }
-        // Humans must not be swept up by the role-suffix rule.
-        for localpart in ["tyreseluo", "alex", "haitang", "coordinator_notes"] {
-            assert!(
-                !is_agent_chat_puppet_localpart(localpart),
-                "{localpart} must NOT be treated as a bot"
-            );
-        }
-    }
-
-    #[test]
-    fn test_agent_reply_kind_split_from_body() {
-        let (kind, rest) = AgentReplyKind::split_from_body("📋 Issue 001 spec ready");
-        assert_eq!(kind, Some(AgentReplyKind::Request));
-        assert_eq!(rest, "Issue 001 spec ready");
-
-        let (kind, rest) = AgentReplyKind::split_from_body("↩️ Review verdict: APPROVE");
-        assert_eq!(kind, Some(AgentReplyKind::Reply));
-        assert_eq!(rest, "Review verdict: APPROVE");
-
-        // No marker → body is returned untouched and no badge is shown.
-        let (kind, rest) = AgentReplyKind::split_from_body("plain agent text");
-        assert_eq!(kind, None);
-        assert_eq!(rest, "plain agent text");
-    }
-
-    #[test]
-    fn test_bot_kind_badge_strips_marker_from_rendered_body() {
-        let state = compute_bot_timeline_render_state("ℹ️ Status update\n\nAll good.", true);
-        assert_eq!(state.kind, Some(AgentReplyKind::Inform));
-        assert!(
-            !state.body.starts_with('ℹ'),
-            "type marker must move to the badge, not stay in the body: {:?}",
-            state.body
-        );
-    }
-
-    #[test]
-    fn test_short_bot_body_is_not_folded() {
-        let state = compute_bot_timeline_render_state("via x\n\nline1\nline2\nline3", true);
-        assert!(
-            state.folded_body.is_none(),
-            "a short reply must render in full with no fold toggle"
-        );
-    }
-
-    #[test]
-    fn test_long_bot_body_folds_to_preview() {
-        let long = (1..=30).map(|i| format!("line {i}")).collect::<Vec<_>>().join("\n");
-        let state = compute_bot_timeline_render_state(&long, true);
-        let folded = state.folded_body.expect("a long reply must fold");
-        assert_eq!(folded.lines().count(), BOT_BODY_FOLD_PREVIEW_LINES);
-        assert!(folded.starts_with("line 1"));
-        // The full text stays available for the expanded state.
-        assert!(state.body.lines().count() > BOT_BODY_FOLD_LINE_THRESHOLD);
-    }
-
-    #[test]
-    fn test_human_messages_are_never_folded_or_badged() {
-        let long = (1..=30).map(|i| format!("line {i}")).collect::<Vec<_>>().join("\n");
-        let state = compute_bot_timeline_render_state(&long, false);
-        assert!(state.folded_body.is_none());
-        assert_eq!(state.kind, None);
-        assert!(!state.show_card);
-    }
-
-    #[test]
-    fn test_bot_metadata_extracted_for_meta_band() {
-        let state = compute_bot_timeline_render_state(
-            "via moonshot@api (kimi-k2.5)\n\n你好！我是 Alex。\n\n_moonshot@api/kimi-k2.5 · 1.2K in · 88 out · 2s_",
-            true,
-        );
-
-        assert!(state.show_metadata_footer);
-        assert_eq!(state.provider.as_deref(), Some("via moonshot@api (kimi-k2.5)"));
-        assert_eq!(
-            state.footer.as_deref(),
-            Some("_moonshot@api/kimi-k2.5 · 1.2K in · 88 out · 2s_"),
-        );
-    }
-
-    #[test]
-    fn test_clipboard_text_strips_bot_scaffolding() {
-        let raw = "via deepseek@api (deepseek-chat)\n\n我运行在 OctOS 平台上。\n\n_deepseek@api/deepseek-chat · 13.3K in · 162 out · 3s_";
-        let cleaned = clipboard_text_for_message_body(raw.to_string(), true);
-
-        assert!(cleaned.contains("我运行在 OctOS 平台上。"));
-        assert!(!cleaned.contains("deepseek@api/deepseek-chat"));
-    }
-
-    #[test]
-    fn test_clipboard_text_verbatim_for_human() {
-        let raw = "via someone@api (model)\nhello there";
-        let unchanged = clipboard_text_for_message_body(raw.to_string(), false);
-
-        assert_eq!(unchanged, raw);
-    }
-
-    #[test]
-    fn test_bot_progress_message_hides_body_card_when_only_metrics_remain() {
-        let state = compute_bot_timeline_render_state(
-            "疯狂输出中\nvia moonshot@api (kimi-k2.5)\n4s",
-            true,
-        );
-
-        assert!(state.show_card);
-        assert!(!state.show_body_card);
-        assert!(state.show_status_strip);
-        assert!(state.show_metadata_footer);
-        assert_eq!(state.footer.as_deref(), Some("4s"));
-    }
-
-    #[test]
-    fn test_bot_timeline_card_body_uses_html_or_plaintext_rendering() {
-        let state = compute_bot_timeline_render_state(
-            "施法中\nvia moonshot@api (kimi-k2.5)\n\n你好！我是 **Alex**",
-            true,
-        );
-
-        let formatted = select_bot_timeline_body_formatted_body(&state, None)
-            .expect("structured bot body should still produce formatted content");
-
-        assert_eq!(formatted.format, MessageFormat::Html);
-        assert!(formatted.body.contains("<strong>Alex</strong>"));
-    }
-
-    #[test]
-    fn test_bot_plain_markdown_body_without_formatted_html_still_renders_as_markdown() {
-        let state = compute_bot_timeline_render_state(
-            "## 标题\n\n```rust\n// 中文注释\nlet answer = 42;\n```",
-            true,
-        );
-
-        let formatted = select_bot_timeline_body_formatted_body(&state, None)
-            .expect("rich markdown bot body should synthesize HTML during streaming");
-
-        assert_eq!(formatted.format, MessageFormat::Html);
-        assert!(formatted.body.contains("<h2>标题</h2>"));
-        assert!(formatted.body.contains("中文注释"));
-    }
-
-    #[test]
-    fn test_bot_timeline_body_prefers_markdown_widget_for_fenced_code_blocks() {
-        let state = compute_bot_timeline_render_state(
-            "## 标题\n\n```rust\nlet answer = 42;\n```\n\n这里是中文说明。",
-            true,
-        );
-
-        assert!(should_render_bot_timeline_body_with_markdown_widget(&state));
-        assert_eq!(
-            bot_timeline_code_block_mode(&state),
-            BotTimelineCodeBlockMode::Highlighted,
-        );
-    }
-
-    #[test]
-    fn test_bot_timeline_body_keeps_html_widget_for_non_code_markdown() {
-        let state = compute_bot_timeline_render_state(
-            "## 标题\n\n这里有 **加粗**，但没有代码块。",
-            true,
-        );
-
-        assert!(!should_render_bot_timeline_body_with_markdown_widget(&state));
-        assert_eq!(
-            bot_timeline_code_block_mode(&state),
-            BotTimelineCodeBlockMode::None,
-        );
-    }
-
-    #[test]
-    fn test_bot_timeline_body_uses_plain_markdown_code_block_for_cjk_code() {
-        let state = compute_bot_timeline_render_state(
-            "```rust\n// 中文注释\nprintln!(\"你好\");\n```",
-            true,
-        );
-
-        assert_eq!(
-            bot_timeline_code_block_mode(&state),
-            BotTimelineCodeBlockMode::Plain,
-        );
-    }
-
-    #[test]
-    fn test_fenced_code_blocks_ignore_cjk_outside_code_block() {
-        let body = "## 标题\n\n```rust\nlet answer = 42;\n```\n\n这里是中文总结。";
-
-        assert!(!fenced_code_blocks_contain_cjk(body));
-    }
-
-    #[test]
-    fn test_streaming_update_requires_content_invalidation_for_new_full_snapshot_text() {
-        let state = StreamingAnimState::new("你好", true);
-
-        assert!(streaming_update_requires_content_invalidation(
-            &state,
-            "## 标题\n\n内容",
-            true,
-            true,
-        ));
-    }
-
-    #[test]
-    fn test_streaming_update_skips_invalidation_when_target_and_mode_are_unchanged() {
-        let mut state = StreamingAnimState::new("## 标题\n\n内容", true);
-        state.set_render_full_target(true);
-
-        assert!(!streaming_update_requires_content_invalidation(
-            &state,
-            "## 标题\n\n内容",
-            true,
-            true,
-        ));
-    }
-
-    #[test]
-    fn test_bot_timeline_card_preserves_reply_preview_and_condensed_layout() {
-        let reply_state = compute_bot_timeline_render_state(
-            "via moonshot@api (kimi-k2.5)\n\n第一条回复",
-            true,
-        );
-        let condensed_state = compute_bot_timeline_render_state(
-            "via moonshot@api (kimi-k2.5)\n\n第二条回复",
-            true,
-        );
-
-        assert!(reply_state.show_card);
-        assert!(condensed_state.show_card);
-        assert!(reply_state.show_metadata_footer);
-        assert!(condensed_state.show_metadata_footer);
-    }
-}
-
-
-#[cfg(test)]
-mod t1_fold_tests {
-    use super::*;
-
-    /// A minified payload is one enormous line: the line count says "nothing to
-    /// fold" while it wraps to fill the viewport.
-    #[test]
-    fn single_enormous_line_folds() {
-        let body = "x".repeat(BOT_BODY_FOLD_CHAR_THRESHOLD + 1);
-        assert_eq!(body.lines().count(), 1, "one line, far below the line threshold");
-        let folded = fold_bot_body_preview(&body).expect("folds on characters");
-        assert!(folded.chars().count() <= BOT_BODY_FOLD_PREVIEW_CHARS + 1);
-        assert!(folded.ends_with('\u{2026}'));
-    }
-
-    /// Three lines of a dump can still be a screenful, so the preview is capped
-    /// by characters as well as by lines.
-    #[test]
-    fn preview_is_capped_by_characters_too() {
-        let body = format!("{}\n{}\n{}\nmore\nmore\nmore\nmore\nmore\nmore",
-            "a".repeat(900), "b".repeat(900), "c".repeat(900));
-        let folded = fold_bot_body_preview(&body).expect("folds");
-        assert!(folded.chars().count() <= BOT_BODY_FOLD_PREVIEW_CHARS + 1);
-    }
-
-    /// Multi-byte text must not panic the character cap.
-    #[test]
-    fn character_cap_is_utf8_safe() {
-        let body = "中文内容描述测试".repeat(400);
-        let folded = fold_bot_body_preview(&body).expect("folds on characters");
-        assert!(folded.chars().count() <= BOT_BODY_FOLD_PREVIEW_CHARS + 1);
-    }
-
-    /// A body that is small on both counts still shows no toggle.
-    #[test]
-    fn short_body_still_does_not_fold() {
-        assert!(fold_bot_body_preview("one\ntwo\nthree").is_none());
-    }
-
-    #[test]
-    fn agent_role_is_derived_from_localpart() {
-        use AgentRole::*;
-        // Real MXIDs from the agent-chat demo: @ac_<team>_<role>
-        assert_eq!(AgentRole::from_localpart("ac_tyrese_coordinator"), Some(Coordinator));
-        assert_eq!(AgentRole::from_localpart("ac_tyrese_implementer"), Some(Implementer));
-        assert_eq!(AgentRole::from_localpart("ac_tyrese_reviewer"), Some(Reviewer));
-        // `final_reviewer` also ends with `reviewer` — it must not be mislabelled.
-        assert_eq!(AgentRole::from_localpart("ac_tyrese_final_reviewer"), Some(FinalReviewer));
-        assert_eq!(AgentRole::from_localpart("ac_wf_final_reviewer"), Some(FinalReviewer));
-        // Bare role names (no team prefix) still resolve.
-        assert_eq!(AgentRole::from_localpart("coordinator"), Some(Coordinator));
-    }
-
-    #[test]
-    fn non_workflow_bots_have_no_role() {
-        assert_eq!(AgentRole::from_localpart("octosbot"), None);
-        assert_eq!(AgentRole::from_localpart("agent-bridge-tyrese"), None);
-        assert_eq!(AgentRole::from_localpart("tyreseluo"), None);
-        // A name that merely contains a role word is not a role.
-        assert_eq!(AgentRole::from_localpart("reviewerbot"), None);
-    }
-
-    /// The real 4:43 message: permalink must leave the body and land in the footer.
-    #[test]
-    fn permalink_is_promoted_out_of_body() {
-        let body = "\u{21a9}\u{fe0f} Saved to docs/weekly/ ... \u{b7} @tyreseluo\n\n@tyreseluo ok:\n\n```\n/Users/x/f.md\n```\n\nmore\nmore\nmore\nmore\nmore\n\n\u{1f517} http://127.0.0.1:8090/msg/msg_0392?view=Frwug96p";
-        let st = compute_bot_timeline_render_state(body, true);
-        assert_eq!(st.permalink.as_deref(), Some("http://127.0.0.1:8090/msg/msg_0392?view=Frwug96p"));
-        assert!(!st.body.contains('\u{1f517}'), "permalink line removed from body");
-        assert!(!st.body.trim_end().ends_with("Frwug96p"), "url gone from body");
-        let folded = st.folded_body.expect("still folds");
-        assert!(!folded.contains('\u{1f517}'), "preview has no permalink");
-    }
-
-    #[test]
-    fn html_anchor_for_permalink_is_stripped() {
-        let html = "<b>hi</b><br><br>body text<br><br><a href=\"http://x/msg/1\">\u{1f517} View formatted</a>";
-        let out = strip_permalink_anchor_from_html(html);
-        assert_eq!(out, "<b>hi</b><br><br>body text");
-    }
-
-    #[test]
-    fn html_without_permalink_anchor_is_untouched() {
-        let html = "<b>hi</b><br><a href=\"http://x\">docs</a>";
-        assert_eq!(strip_permalink_anchor_from_html(html), html);
-    }
-
-    #[test]
-    fn body_without_permalink_is_untouched() {
-        let (body, link) = split_bot_permalink("just text\nsecond line");
-        assert_eq!(body, "just text\nsecond line");
-        assert!(link.is_none());
-    }
-
-    /// The screenshot-3 regression: a long body containing a fenced code block
-    /// must fold to a preview that is itself renderable by a single widget.
-    #[test]
-    fn folded_preview_is_fence_balanced_and_short() {
-        let body = "↩️ Summary line here\n\n@user ok:\n\n```\n/Users/x/file.md\n```\n\nmore text\nand more\nand more\nand more\nand more";
-        let st = compute_bot_timeline_render_state(body, true);
-        assert_eq!(st.kind, Some(AgentReplyKind::Reply), "type marker parsed");
-        assert!(!st.body.starts_with('↩'), "marker stripped from body");
-        let folded = st.folded_body.expect("long body folds");
-        assert!(folded.lines().count() <= BOT_BODY_FOLD_PREVIEW_LINES + 1);
-        assert_eq!(folded.matches("```").count() % 2, 0, "fences balanced");
-    }
-
-    #[test]
-    fn short_body_does_not_fold() {
-        let st = compute_bot_timeline_render_state("ℹ️ short\n\none line", true);
-        assert_eq!(st.kind, Some(AgentReplyKind::Inform));
-        assert!(st.folded_body.is_none(), "short body must not show a toggle");
-    }
-
-    #[test]
-    fn human_message_is_untouched() {
-        let body = "📋 not a bot\nline\nline\nline\nline\nline\nline\nline\nline\nline";
-        let st = compute_bot_timeline_render_state(body, false);
-        assert!(st.kind.is_none(), "no badge for human senders");
-        assert!(st.folded_body.is_none(), "no folding for human senders");
-        assert_eq!(st.body, body, "human body verbatim");
-    }
 }
